refactor(cli): decompose image/mod.rs shared OCI core into submodules
Split the ~2900-line shared OCI core into oci_types (data types),
cache (index + file I/O + ls/inspect/rm bodies), trust_policy (the
frozen registry-allowlist/cosign enforcement), pull_core (registry
resolve/pull), materialize (rootfs injection + runtime-tag sealing),
and ingest (local archive/stdin/rootfs-dir sources). mod.rs is now a
thin dispatch + re-export shell that keeps the six existing sibling
verb modules' `super::<name>` calls and vm/exec.rs's
`super::super::image::<name>` reach-in resolving unchanged. Pure
structural move: no behavior, output, or trust-decision logic changed.

check-guest-binary-lists' find() scan follows embedded_guest_binaries
to its new home in materialize.rs.
