+++ crates/mvm-cli/src/commands/image/mod.rs
@@ -1,39 +1,40 @@
 //! `mvmctl image` - pull, inspect, and prune the local OCI image cache.
 
-use std::collections::{BTreeMap, BTreeSet};
-use std::fs;
-use std::io::{BufReader, Cursor, Read};
-use std::path::{Component, Path, PathBuf};
+use std::path::PathBuf;
 
-use anyhow::{Context, Result, bail};
+use anyhow::{Result, bail};
 use clap::{Args as ClapArgs, Subcommand};
-use flate2::read::GzDecoder;
-use mvm_build::oci_runtime_inject::OciEntrypointConfig;
-use mvm_build::rootfs::MaterializeExt4Input;
-use mvm_fs::oci::{
-    ImageReference, LayerDescriptor, LayerFetchOptions, LinuxPlatform, OciLayerFetcher,
-    OciManifestFetcher, RegistryAuthConfig, UnpackOptions, read_oci_archive, unpack_layer,
-    verify_sha256_digest,
-};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
 
-use mvm_core::domain::manifest::canonical_key_for_path;
+// Kept here (rather than folded into a submodule's own imports) so the
+// unmoved `trust` sibling's `use super::*;` glob keeps resolving these
+// without touching its call sites.
+use mvm_fs::oci::{ImageReference, RegistryAuthConfig};
+
 use mvm_core::user_config::MvmConfig;
 
 use super::Cli;
-use super::shared::human_bytes;
 
+mod cache;
+mod ingest;
 mod inspect;
 mod ls;
+mod materialize;
+mod oci_types;
 mod pull;
+mod pull_core;
 mod rm;
 mod source;
 mod trust;
+mod trust_policy;
 
-use trust::{CosignCommandVerifier, CosignVerifier, default_require_signatures, registry_auth_for};
-
-const INDEX_FILE: &str = "index.json";
+// Re-exports preserve two frozen call-site shapes without editing them: the
+// sibling verb modules' `super::<name>` calls, and
+// `crate::commands::vm::exec`'s `super::super::image::<name>` reach-in.
+use cache::{inspect_image, list_rows, remove_image, render_inspect, render_list};
+use oci_types::CosignIdentity;
+pub(in crate::commands) use oci_types::ResolvedOciRunImage;
+use pull_core::pull_image_with_trust;
+pub(in crate::commands) use pull_core::resolve_or_pull_run_image;
 
 #[derive(ClapArgs, Debug, Clone)]
 pub(in crate::commands) struct Args {
@@ -75,248 +76,6 @@ pub(in crate::commands) enum ImageAction {
     },
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct OciCacheIndex {
-    #[serde(default = "schema_version")]
-    schema_version: u32,
-    #[serde(default)]
-    images: Vec<CachedOciImage>,
-}
-
-impl Default for OciCacheIndex {
-    fn default() -> Self {
-        // Hand-written, NOT derived: a derived `Default` sets `schema_version`
-        // to 0 (the u32 default), which `save_index` then persists and the next
-        // `load_index` rejects as unsupported. The `#[serde(default)]` only
-        // fills a *missing* field on deserialize — it does not feed `Default`.
-        Self {
-            schema_version: schema_version(),
-            images: Vec::new(),
-        }
-    }
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct OciImageConfigInner {
-    #[serde(default, rename = "Entrypoint")]
-    entrypoint: Option<Vec<String>>,
-    #[serde(default, rename = "Cmd")]
-    cmd: Option<Vec<String>>,
-    #[serde(default, rename = "Env")]
-    env: Vec<String>,
-    #[serde(default, rename = "WorkingDir")]
-    working_dir: Option<String>,
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct OciImageConfig {
-    #[serde(default, rename = "config")]
-    config: OciImageConfigInner,
-}
-
-fn schema_version() -> u32 {
-    1
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(in crate::commands) struct CachedOciImage {
-    reference: String,
-    registry: String,
-    repository: String,
-    tag: Option<String>,
-    resolved_digest: String,
-    fetched_at: String,
-    manifest_path: String,
-    #[serde(default)]
-    config_path: Option<String>,
-    #[serde(default)]
-    rootfs_path: Option<String>,
-    /// Runtime identity (crate version + epoch) of the guest agent/netinit
-    /// baked into `rootfs_path`. A mismatch with the running mvmctl forces
-    /// re-materialization so a stale agent — e.g. one predating the interactive
-    /// exec handler — is never reused. Absent on entries written before the
-    /// tag existed, which therefore read as stale and re-materialize.
-    #[serde(default)]
-    runtime_tag: Option<String>,
-    #[serde(default)]
-    claims_path: Option<String>,
-    #[serde(default)]
-    layers: Vec<CachedOciLayer>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(in crate::commands) struct ResolvedOciRunImage {
-    pub reference: String,
-    pub resolved_digest: String,
-    pub rootfs_path: PathBuf,
-    /// The prepared rootfs-only OCI tree, when it is present on disk — the
-    /// source a virtiofs-root dev boot serves directly. This is distinct from
-    /// the raw unpacked layer tree so the virtiofs `RootfsOnly` contract does
-    /// not share one injected runtime layout with the block-backed ext4 image.
-    pub unpacked_root: Option<PathBuf>,
-    pub pulled: bool,
-    pub provenance: OciProvenance,
-    pub auth_source: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(in crate::commands) struct OciProvenance {
-    pub schema_version: u32,
-    pub source: String,
-    pub supplied_reference: String,
-    pub canonical_reference: String,
-    pub registry: String,
-    pub repository: String,
-    pub tag: Option<String>,
-    pub resolved_digest: String,
-    pub layer_digests: Vec<String>,
-    pub trust_policy: String,
-    pub verification_status: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct CachedOciLayer {
-    digest: String,
-    #[serde(default)]
-    size_bytes: u64,
-    #[serde(default)]
-    path: Option<String>,
-}
-
-impl CachedOciImage {
-    pub(super) fn provenance(
-        &self,
-        source: &str,
-        supplied_reference: &str,
-        trust: &OciTrustDecision,
-    ) -> OciProvenance {
-        OciProvenance {
-            schema_version: 1,
-            source: source.to_string(),
-            supplied_reference: supplied_reference.to_string(),
-            canonical_reference: self.reference.clone(),
-            registry: self.registry.clone(),
-            repository: self.repository.clone(),
-            tag: self.tag.clone(),
-            resolved_digest: self.resolved_digest.clone(),
-            layer_digests: self
-                .layers
-                .iter()
-                .map(|layer| layer.digest.clone())
-                .collect(),
-            trust_policy: trust.trust_policy.clone(),
-            verification_status: trust.verification_status.clone(),
-        }
-    }
-}
-
-impl OciProvenance {
-    pub(in crate::commands) fn audit_labels(&self) -> Vec<(String, String)> {
-        vec![
-            ("oci_source".to_string(), self.source.clone()),
-            (
-                "oci_supplied_reference".to_string(),
-                self.supplied_reference.clone(),
-            ),
-            (
-                "oci_canonical_reference".to_string(),
-                self.canonical_reference.clone(),
-            ),
-            ("oci_registry".to_string(), self.registry.clone()),
-            ("oci_repository".to_string(), self.repository.clone()),
-            (
-                "oci_resolved_digest".to_string(),
-                self.resolved_digest.clone(),
-            ),
-            (
-                "oci_layer_digests".to_string(),
-                self.layer_digests.join(","),
-            ),
-            ("oci_trust_policy".to_string(), self.trust_policy.clone()),
-            (
-                "oci_verification_status".to_string(),
-                self.verification_status.clone(),
-            ),
-        ]
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(super) struct OciTrustDecision {
-    trust_policy: String,
-    verification_status: String,
-}
-
-impl OciTrustDecision {
-    fn dev_digest_only(image_ref: &ImageReference) -> Self {
-        let trust_policy = if image_ref.is_digest_pinned() {
-            "digest-pinned"
-        } else {
-            "mutable-reference-resolved-to-digest"
-        };
-        Self {
-            trust_policy: trust_policy.to_string(),
-            verification_status: "digest-verified-signature-not-required".to_string(),
-        }
-    }
-
-    fn cosign_verified(identity: &CosignIdentity) -> Self {
-        Self {
-            trust_policy: "prod-cosign-required".to_string(),
-            verification_status: format!(
-                "cosign-verified identity={} issuer={}",
-                identity.certificate_identity, identity.certificate_oidc_issuer
-            ),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-struct OciRegistryPolicy {
-    #[serde(default)]
-    allowed_registries: Vec<String>,
-    #[serde(default = "default_require_signatures")]
-    require_signatures: bool,
-    #[serde(default)]
-    cosign: Vec<CosignIdentity>,
-}
-
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-struct CosignIdentity {
-    certificate_identity: String,
-    certificate_oidc_issuer: String,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-pub(super) struct ImageListRow {
-    reference: String,
-    registry: String,
-    repository: String,
-    tag: Option<String>,
-    resolved_digest: String,
-    fetched_at: String,
-    size_bytes: u64,
-    layers: usize,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct InspectOutput {
-    image: CachedOciImage,
-    size_bytes: u64,
-    manifest: Option<Value>,
-    config: Option<Value>,
-    claims: Option<Value>,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub(super) struct RemoveOutcome {
-    pub(super) reference: String,
-    pub(super) removed_files: usize,
-    pub(super) freed_bytes: u64,
-}
-
 pub(in crate::commands) fn run(_cli: &Cli, args: Args, _cfg: &MvmConfig) -> Result<()> {
     let cache_root = oci_cache_root();
     match args.action {
@@ -329,2558 +88,4 @@ pub(in crate::commands) fn run(_cli: &Cli, args: Args, _cfg: &MvmConfig) -> Resu
 
 pub(in crate::commands) fn oci_cache_root() -> PathBuf {
     PathBuf::from(mvm_core::config::mvm_cache_dir()).join("oci")
-}
-
-/// Bump when the injected OCI runtime (guest agent, netinit, or `/init`
-/// script) changes such that already-materialized rootfs images must be
-/// re-sealed. Folded with the crate version into [`oci_runtime_tag`]; a
-/// stale rootfs then re-materializes instead of booting an outdated agent.
-/// Epoch 1 introduces the interactive exec handler in the OCI agent. Epoch 2 adds
-/// the detached-workload handler — a rootfs sealed with an older agent would
-/// reject the run-detached request, so it must re-materialize. Epoch 3 moves
-/// `mvm-egress-client` onto the runtime overlay contract: cached injected
-/// rootfs images that still bake the old helper path must be re-materialized so
-/// block-backed boots pick up the overlay-first layout. Epoch 4 rolls the
-/// required-overlay `/init` contract fix that initializes the vsock-egress
-/// flag safely under `set -u`; stale cached rootfs images otherwise keep
-/// booting the pre-fix PID-1 script and panic before the agent binds. Epoch 5
-/// teaches that same `/init` to skip re-mounting `/mvm/runtime` when
-/// `mvm-verity-init` already mounted the verity-protected overlay there; stale
-/// cached runtime-lean rootfs entries otherwise panic after the verity handoff.
-const OCI_RUNTIME_EPOCH: u32 = 6;
-
-/// Identity of the guest runtime the running mvmctl bakes into an OCI rootfs.
-/// Cheap and build-free (no agent cross-compile) so it can gate the cache-hit
-/// path without forcing a rebuild on hosts that only have a warm rootfs cache.
-fn oci_runtime_tag() -> String {
-    use sha2::{Digest, Sha256};
-
-    let mut hasher = Sha256::new();
-    hasher.update(b"epoch=");
-    hasher.update(OCI_RUNTIME_EPOCH.to_string().as_bytes());
-    hasher.update(b"\n");
-    for name in [
-        "mvm-oci-init",
-        "mvm-oci-entrypoint",
-        "mvm-guest-agent",
-        "mvm-guest-netinit",
-        "mvm-guest-netd",
-        "mvm-egress-client",
-        "mvm-verity-init",
-    ] {
-        hasher.update(name.as_bytes());
-        hasher.update(b"=");
-        let sha = crate::host_binaries::embedded::EMBEDDED
-            .iter()
-            .find(|bin| bin.name == name)
-            .map(|bin| bin.sha256_hex)
-            .unwrap_or("missing");
-        hasher.update(sha.as_bytes());
-        hasher.update(b"\n");
-    }
-    let digest = hex::encode(hasher.finalize());
-    format!("{}-guest-{}", env!("CARGO_PKG_VERSION"), &digest[..16])
-}
-
-fn oci_tree_key(identity: &str) -> String {
-    sha256_hex(identity).unwrap_or_else(|_| identity.replace(['/', ':'], "-"))
-}
-
-fn prepared_virtiofs_root(cache_root: &Path, identity: &str, runtime_tag: &str) -> PathBuf {
-    cache_root
-        .join("prepared-roots")
-        .join(format!("{}-{runtime_tag}", oci_tree_key(identity)))
-        .join("rootfs-only")
-}
-
-fn oci_entrypoint_from_config_bytes(bytes: &[u8]) -> Result<Option<OciEntrypointConfig>> {
-    let config: OciImageConfig = serde_json::from_slice(bytes).context("parse OCI image config")?;
-    let mut argv = config.config.entrypoint.unwrap_or_default();
-    argv.extend(config.config.cmd.unwrap_or_default());
-    if argv.is_empty() {
-        return Ok(None);
-    }
-    let env = config.config.env;
-    let working_dir = config.config.working_dir.filter(|dir| !dir.is_empty());
-    Ok(Some(OciEntrypointConfig {
-        argv,
-        env,
-        working_dir,
-    }))
-}
-
-fn oci_entrypoint_from_cache_path(
-    cache_root: &Path,
-    config_path: Option<&str>,
-) -> Result<Option<OciEntrypointConfig>> {
-    let Some(config_path) = config_path else {
-        return Ok(None);
-    };
-    let path = safe_cache_path(cache_root, config_path)?;
-    let bytes = fs::read(&path).with_context(|| format!("read OCI config {}", path.display()))?;
-    oci_entrypoint_from_config_bytes(&bytes)
-}
-
-/// Whether a cached image's materialized rootfs can be booted as-is: it must
-/// exist and carry the current runtime tag. A `None` tag (pre-tag entry) or a
-/// mismatch means the baked agent may be outdated, so the rootfs is stale.
-fn cached_rootfs_is_current(cached: &CachedOciImage, runtime_tag: &str) -> bool {
-    cached.rootfs_path.is_some() && cached.runtime_tag.as_deref() == Some(runtime_tag)
-}
-
-fn rootfs_verity_sidecars_present(rootfs_path: &Path) -> bool {
-    let Some(parent) = rootfs_path.parent() else {
-        return false;
-    };
-    parent.join("rootfs.verity").is_file() && parent.join("rootfs.roothash").is_file()
-}
-
-fn ensure_rootfs_verity_sidecars(
-    rootfs_path: &Path,
-    image_reference: &str,
-    unpacked_root: Option<&Path>,
-) -> Result<()> {
-    if rootfs_verity_sidecars_present(rootfs_path) {
-        return Ok(());
-    }
-    let unpacked_note = unpacked_root
-        .map(|path| format!("cached unpacked tree: {}", path.display()))
-        .unwrap_or_else(|| "cached unpacked tree: unavailable".to_string());
-    bail!(
-        "cached OCI image {} is missing sealed block-root sidecars beside {} \
-         after materialization; required files rootfs.verity and rootfs.roothash were not produced ({})",
-        image_reference,
-        rootfs_path.display(),
-        unpacked_note
-    );
-}
-
-pub(in crate::commands) fn resolve_or_pull_run_image(
-    cache_root: &Path,
-    reference: &str,
-    prod: bool,
-) -> Result<ResolvedOciRunImage> {
-    resolve_or_pull_run_image_with(cache_root, reference, prod, inject_runtime_and_materialize)
-}
-
-type RuntimeMaterializer =
-    fn(&Path, &Path, &Path, &str, Option<&OciEntrypointConfig>, bool) -> Result<()>;
-
-fn resolve_or_pull_run_image_with(
-    cache_root: &Path,
-    reference: &str,
-    prod: bool,
-    materialize: RuntimeMaterializer,
-) -> Result<ResolvedOciRunImage> {
-    // Rootfs materialization can fall back to a builder VM when the in-process
-    // ext4 writer cannot faithfully emit a tree. `mvmctl image pull` reaps
-    // previous helper processes at startup; the run-image path is the other
-    // place a builder VM can spawn, so give it the same sweep before we might add one.
-    crate::commands::env::builder_vm::sweep_orphaned_vm_helpers_on_startup();
-
-    // Local sources route to their own ingest; a registry reference falls
-    // through to the cache-or-pull path below.
-    match source::ImageSource::classify(reference)? {
-        source::ImageSource::OciArchive(path) => {
-            return ingest_local_archive(cache_root, &path, reference, prod);
-        }
-        source::ImageSource::Stdin => {
-            return ingest_stdin_archive(cache_root, reference, prod);
-        }
-        source::ImageSource::RootfsDir(path) => {
-            return ingest_rootfs_dir(cache_root, &path, reference, prod);
-        }
-        source::ImageSource::Registry(_) => {}
-    }
-    let image_ref: ImageReference = reference.parse()?;
-    if prod && !image_ref.is_digest_pinned() {
-        bail!("mvmctl run --image --prod requires a digest-pinned reference");
-    }
-    let canonical = image_ref.canonical();
-    let runtime_tag = oci_runtime_tag();
-    let (image, pulled, trust_from_pull, auth_source_from_pull) = match load_index(cache_root)
-        .ok()
-        .and_then(|index| find_image(&index, &canonical).cloned())
-    {
-        Some(cached) if cached_rootfs_is_current(&cached, &runtime_tag) => {
-            (cached, false, None, None)
-        }
-        Some(cached) => {
-            match rematerialize_cached_image(cache_root, cached, &runtime_tag, materialize, prod)? {
-                Some(repaired) => (repaired, false, None, None),
-                None => {
-                    let (cached, trust, auth_source) =
-                        pull_image_ref(cache_root, image_ref.clone(), reference, prod)?;
-                    (cached, true, Some(trust), Some(auth_source))
-                }
-            }
-        }
-        _ => {
-            let (cached, trust, auth_source) =
-                pull_image_ref(cache_root, image_ref.clone(), reference, prod)?;
-            (cached, true, Some(trust), Some(auth_source))
-        }
-    };
-    let Some(rootfs_relative) = image.rootfs_path.as_deref() else {
-        bail!(
-            "cached OCI image {} has no materialized rootfs; run `mvmctl image pull {}` first",
-            image.reference,
-            image.reference
-        );
-    };
-    let rootfs_path = safe_cache_path(cache_root, rootfs_relative)?;
-    let mut rematerialized_from = None;
-    if !rootfs_path.is_file() || !rootfs_verity_sidecars_present(&rootfs_path) {
-        // Self-heal a cache whose index still records a materialized rootfs but
-        // whose sealed block-root artifacts have since drifted. That covers a
-        // vanished `rootfs.ext4` (interrupted prune / manual delete) and older
-        // ext4-only cache entries that predate the current verity-backed OCI
-        // materializer. If the unpacked layer tree survives, re-run the same
-        // seal `image pull` performs — network-free, from the cached layers —
-        // rather than failing the run. Only when the unpacked tree is gone too
-        // is this a genuine cache loss the user must re-pull.
-        match unpacked_dir_if_present(cache_root, &image.resolved_digest) {
-            Some(unpacked_root) => {
-                materialize(
-                    cache_root,
-                    &unpacked_root,
-                    &rootfs_path,
-                    &image.reference,
-                    oci_entrypoint_from_cache_path(cache_root, image.config_path.as_deref())?
-                        .as_ref(),
-                    prod,
-                )
-                .with_context(|| {
-                    format!(
-                        "re-materializing cached OCI rootfs artifacts for {} from {}",
-                        image.reference,
-                        unpacked_root.display()
-                    )
-                })?;
-                rematerialized_from = Some(unpacked_root);
-            }
-            None => bail!(
-                "cached OCI image {} is missing sealed rootfs artifacts beside {} and its unpacked \
-                 layers are gone; run `mvmctl image pull {}` to re-fetch",
-                image.reference,
-                rootfs_path.display(),
-                image.reference
-            ),
-        }
-    }
-    ensure_rootfs_verity_sidecars(
-        &rootfs_path,
-        &image.reference,
-        rematerialized_from.as_deref(),
-    )?;
-    let trust = match trust_from_pull {
-        Some(trust) => trust,
-        None => trust_decision_for_cached_image(&image_ref, &image, prod, &CosignCommandVerifier)?,
-    };
-    let unpacked_root = unpacked_dir_if_present(cache_root, &image.resolved_digest)
-        .map(|raw| prepare_rootfs_only_tree(cache_root, &raw, &image.resolved_digest))
-        .transpose()?;
-    Ok(ResolvedOciRunImage {
-        provenance: image.provenance("run_image", reference, &trust),
-        reference: image.reference,
-        resolved_digest: image.resolved_digest,
-        rootfs_path,
-        unpacked_root,
-        pulled,
-        auth_source: auth_source_from_pull,
-    })
-}
-
-fn rematerialize_cached_image(
-    cache_root: &Path,
-    mut image: CachedOciImage,
-    runtime_tag: &str,
-    materialize: RuntimeMaterializer,
-    prod: bool,
-) -> Result<Option<CachedOciImage>> {
-    let Some(unpacked_root) = unpacked_dir_if_present(cache_root, &image.resolved_digest) else {
-        return Ok(None);
-    };
-    let rootfs_path = match (
-        image.rootfs_path.as_deref(),
-        image.runtime_tag.as_deref() == Some(runtime_tag),
-    ) {
-        (Some(path), true) => path.to_string(),
-        _ => format!(
-            "rootfs/{}-{runtime_tag}/rootfs.ext4",
-            sha256_hex(&image.resolved_digest)?
-        ),
-    };
-    let rootfs_abs = safe_cache_path(cache_root, &rootfs_path)?;
-    if !rootfs_abs.is_file() {
-        materialize(
-            cache_root,
-            &unpacked_root,
-            &rootfs_abs,
-            &image.reference,
-            oci_entrypoint_from_cache_path(cache_root, image.config_path.as_deref())?.as_ref(),
-            prod,
-        )
-        .with_context(|| {
-            format!(
-                "re-materializing cached OCI image {} from {}",
-                image.reference,
-                unpacked_root.display()
-            )
-        })?;
-    }
-    image.rootfs_path = Some(rootfs_path);
-    image.runtime_tag = Some(runtime_tag.to_string());
-    let mut index = load_index(cache_root)?;
-    upsert_image(&mut index, image.clone());
-    save_index(cache_root, &index)?;
-    Ok(Some(image))
-}
-
-/// Inject the mvm guest runtime into `unpacked_root`, seal it into
-/// `rootfs_abs` as ext4, and drop the overlay-aware sidecar beside it.
-///
-/// Every OCI run source funnels through here. An arbitrary OCI image has
-/// no mvm agent, so without the injection `run --image` boots a guest
-/// with no vsock control plane and times out at `wait_for_agent`. The
-/// injected `/init` + baked agent + `/mvm/runtime` mount point make the
-/// rootfs genuinely overlay-aware, so the `for_oci_run` sidecar admits
-/// honestly through `admit_overlay_aware` without weakening the gate.
-/// Materialize the run-path rootfs image from an unpacked tree.
-///
-/// Default: the pure in-process `mvm-ext4` writer — no builder VM, no `mkfs`,
-/// no subprocess. `MVM_MATERIALIZE_BUILDER_VM` (any value) routes back through
-/// the builder-VM `mkfs` path for parity / debugging. Both paths emit
-/// `rootfs.verity` / `rootfs.roothash`, so OCI block roots boot sealed no
-/// matter which materializer produced them.
-fn materialize_run_rootfs(input: &MaterializeExt4Input) -> Result<()> {
-    mvm_build::run_image::materialize_run_rootfs(input)
-}
-
-fn inject_runtime_and_materialize(
-    cache_root: &Path,
-    unpacked_root: &Path,
-    rootfs_abs: &Path,
-    image_label: &str,
-    entrypoint: Option<&OciEntrypointConfig>,
-    sealed: bool,
-) -> Result<()> {
-    mvm_build::run_image::inject_and_materialize(
-        mvm_build::run_image::InjectAndMaterializeRequest::builder(
-            cache_root,
-            unpacked_root,
-            rootfs_abs,
-            image_label,
-        )
-        .profile(mvm_build::oci_runtime_inject::RuntimeInjectionProfile::RuntimeLean)
-        .entrypoint(entrypoint)
-        .prebuilt(embedded_guest_binaries())
-        .sealed(sealed)
-        .build(),
-    )
-}
-
-fn prepare_rootfs_only_tree(
-    cache_root: &Path,
-    raw_unpacked_root: &Path,
-    identity: &str,
-) -> Result<PathBuf> {
-    let prepared_root = prepared_virtiofs_root(cache_root, identity, &oci_runtime_tag());
-    if prepared_root.exists() {
-        return Ok(prepared_root);
-    }
-    if let Some(parent) = prepared_root.parent() {
-        fs::create_dir_all(parent).with_context(|| format!("create {}", parent.display()))?;
-    }
-    copy_tree(raw_unpacked_root, &prepared_root).with_context(|| {
-        format!(
-            "copy raw OCI tree {} -> {}",
-            raw_unpacked_root.display(),
-            prepared_root.display()
-        )
-    })?;
-    let bins = mvm_build::run_image::resolve_guest_binaries(cache_root, embedded_guest_binaries())
-        .context("resolve guest binaries for rootfs-only OCI tree")?;
-    mvm_build::oci_runtime_inject::inject_mvm_runtime(
-        &prepared_root,
-        &bins,
-        None,
-        false,
-        mvm_build::oci_runtime_inject::RuntimeInjectionProfile::RootfsOnly,
-    )
-    .with_context(|| {
-        format!(
-            "inject rootfs-only runtime into {}",
-            prepared_root.display()
-        )
-    })?;
-    Ok(prepared_root)
-}
-
-fn materialize_overlay_lean_rootfs(
-    cache_root: &Path,
-    raw_unpacked_root: &Path,
-    rootfs_abs: &Path,
-    image_label: &str,
-) -> Result<()> {
-    let staging_root = rootfs_abs.with_extension("staging");
-    if staging_root.exists() {
-        fs::remove_dir_all(&staging_root)
-            .with_context(|| format!("remove stale staging dir {}", staging_root.display()))?;
-    }
-    copy_tree(raw_unpacked_root, &staging_root).with_context(|| {
-        format!(
-            "copy raw OCI tree {} -> {}",
-            raw_unpacked_root.display(),
-            staging_root.display()
-        )
-    })?;
-    let result = inject_runtime_and_materialize(
-        cache_root,
-        &staging_root,
-        rootfs_abs,
-        image_label,
-        None,
-        false,
-    );
-    let cleanup = fs::remove_dir_all(&staging_root);
-    if let Err(err) = cleanup
-        && staging_root.exists()
-    {
-        tracing::warn!(
-            error = %err,
-            staging = %staging_root.display(),
-            "failed to remove OCI runtime-lean staging tree"
-        );
-    }
-    result
-}
-
-#[cfg(unix)]
-fn copy_tree(src: &Path, dst: &Path) -> std::io::Result<()> {
-    let mut stack = vec![(src.to_path_buf(), dst.to_path_buf())];
-    while let Some((from_dir, to_dir)) = stack.pop() {
-        fs::create_dir_all(&to_dir)?;
-        for entry in fs::read_dir(&from_dir)? {
-            let entry = entry?;
-            let ft = entry.file_type()?;
-            let from = entry.path();
-            let to = to_dir.join(entry.file_name());
-            if ft.is_symlink() {
-                let target = fs::read_link(&from)?;
-                let _ = fs::remove_file(&to);
-                std::os::unix::fs::symlink(&target, &to)?;
-            } else if ft.is_dir() {
-                stack.push((from, to));
-            } else {
-                fs::copy(&from, &to)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-/// The guest-agent binaries embedded in this mvmctl at build time (build.rs).
-/// `None` only if a required embedded binary is unexpectedly missing.
-fn embedded_guest_binaries() -> Option<mvm_build::run_image::PrebuiltGuestBinaries<'static>> {
-    let find = |name: &str| {
-        crate::host_binaries::embedded::EMBEDDED
-            .iter()
-            .find(|b| b.name == name)
-            .map(|b| b.bytes)
-            .filter(|b| !b.is_empty())
-    };
-    Some(mvm_build::run_image::PrebuiltGuestBinaries {
-        oci_init: find("mvm-oci-init")?,
-        agent: find("mvm-guest-agent")?,
-        netinit: find("mvm-guest-netinit")?,
-        netd: find("mvm-guest-netd")?,
-        egress_client: find("mvm-egress-client")?,
-        entrypoint_runner: find("mvm-oci-entrypoint")?,
-        verity_init: find("mvm-verity-init")?,
-    })
-}
-
-/// Ingest a local OCI image-layout archive (`oci-archive:<path>`) into a
-/// bootable rootfs. Mirrors the registry pull's unpack → ext4 path with no
-/// network: `read_oci_archive` digest-verifies every blob, then the same
-/// hardened `unpack_layer` extracts each layer. Dev-only for now — `--prod`
-/// still requires a registry reference + cosign (a local archive carries no
-/// registry policy or signature to check yet). Re-ingests each run rather than
-/// caching by reference (the archive has no stable registry identity).
-fn ingest_local_archive(
-    cache_root: &Path,
-    archive_path: &Path,
-    supplied_reference: &str,
-    prod: bool,
-) -> Result<ResolvedOciRunImage> {
-    if prod {
-        bail!(
-            "`run --image --prod` requires a registry reference with cosign; \
-             local OCI archives are dev-only for now"
-        );
-    }
-    let file = fs::File::open(archive_path)
-        .with_context(|| format!("open OCI archive {}", archive_path.display()))?;
-    ingest_archive_from_reader(
-        cache_root,
-        BufReader::new(file),
-        supplied_reference,
-        "local_archive",
-        &format!("OCI archive {}", archive_path.display()),
-    )
-}
-
-/// Ingest an OCI image-layout archive streamed on stdin (`run --image -`). Same
-/// digest-verified path as a file archive, for CI/agent pipelines that pipe an
-/// archive in. Dev-only (see [`ingest_local_archive`]).
-fn ingest_stdin_archive(
-    cache_root: &Path,
-    supplied_reference: &str,
-    prod: bool,
-) -> Result<ResolvedOciRunImage> {
-    if prod {
-        bail!(
-            "`run --image --prod` requires a registry reference with cosign; \
-             stdin OCI archives are dev-only for now"
-        );
-    }
-    let stdin = std::io::stdin();
-    ingest_archive_from_reader(
-        cache_root,
-        stdin.lock(),
-        supplied_reference,
-        "stdin_archive",
-        "OCI archive on stdin",
-    )
-}
-
-/// Shared archive-ingest core: digest-verify the archive from `reader`, unpack
-/// every layer through the hardened `unpack_layer`, materialize the ext4
-/// rootfs, and build a `ResolvedOciRunImage` with local-source provenance. The
-/// file and stdin callers differ only in where the bytes come from and the
-/// provenance `source` label.
-fn ingest_archive_from_reader<R: Read>(
-    cache_root: &Path,
-    reader: R,
-    supplied_reference: &str,
-    source_label: &str,
-    source_descr: &str,
-) -> Result<ResolvedOciRunImage> {
-    let image = read_oci_archive(reader, &LinuxPlatform::for_current_arch())
-        .with_context(|| format!("read {source_descr}"))?;
-
-    let manifest_hex = sha256_hex(&image.manifest_digest)?;
-    let unpacked_root = cache_root.join("unpacked").join(&manifest_hex);
-    if unpacked_root.exists() {
-        fs::remove_dir_all(&unpacked_root)
-            .with_context(|| format!("remove stale unpacked root {}", unpacked_root.display()))?;
-    }
-    fs::create_dir_all(&unpacked_root)
-        .with_context(|| format!("create {}", unpacked_root.display()))?;
-    for layer in &image.layers {
-        unpack_layer_bytes(&layer.descriptor, &layer.bytes, &unpacked_root)
-            .with_context(|| format!("unpack layer {}", layer.descriptor.digest))?;
-    }
-
-    let rootfs_rel = format!("rootfs/{manifest_hex}-{}/rootfs.ext4", oci_runtime_tag());
-    let rootfs_abs = cache_root.join(&rootfs_rel);
-    let rootfs_only_tree =
-        prepare_rootfs_only_tree(cache_root, &unpacked_root, &image.manifest_digest)?;
-    materialize_overlay_lean_rootfs(
-        cache_root,
-        &unpacked_root,
-        &rootfs_abs,
-        &image.manifest_digest,
-    )?;
-
-    let provenance = OciProvenance {
-        schema_version: 1,
-        source: source_label.to_string(),
-        supplied_reference: supplied_reference.to_string(),
-        canonical_reference: image.manifest_digest.clone(),
-        registry: String::new(),
-        repository: String::new(),
-        tag: None,
-        resolved_digest: image.manifest_digest.clone(),
-        layer_digests: image
-            .layers
-            .iter()
-            .map(|l| l.descriptor.digest.clone())
-            .collect(),
-        trust_policy: "local-archive-digest-verified".to_string(),
-        verification_status: "content-digest-verified (dev)".to_string(),
-    };
-    Ok(ResolvedOciRunImage {
-        reference: image.manifest_digest.clone(),
-        resolved_digest: image.manifest_digest,
-        rootfs_path: rootfs_abs,
-        unpacked_root: Some(rootfs_only_tree),
-        pulled: true,
-        provenance,
-        auth_source: None,
-    })
-}
-
-/// Ingest an already-unpacked rootfs directory (`run --image rootfs-dir:<path>`)
-/// straight into an ext4 rootfs — no layers to unpack. A bare tree carries no
-/// OCI manifest/config, so there is no digest or signature to verify; it is
-/// therefore **dev-only** (`--prod` is refused) and the wrong-architecture
-/// guard falls back to probing a representative ELF binary's machine type
-/// against the host. `materialize_ext4` reads the directory read-only.
-fn ingest_rootfs_dir(
-    cache_root: &Path,
-    rootfs_dir: &Path,
-    supplied_reference: &str,
-    prod: bool,
-) -> Result<ResolvedOciRunImage> {
-    if prod {
-        bail!(
-            "`run --image --prod` requires a registry reference with cosign; an \
-             unpacked rootfs directory carries no provenance to verify"
-        );
-    }
-    if !rootfs_dir.is_dir() {
-        bail!(
-            "rootfs directory not found or not a directory: {}",
-            rootfs_dir.display()
-        );
-    }
-    // A bare tree has no manifest architecture, so probe a representative ELF
-    // binary and reject a cross-arch tree before paying for materialization.
-    if let (Some(found), Some(host)) = (probe_rootfs_elf_machine(rootfs_dir), host_elf_machine())
-        && found != host
-    {
-        bail!(
-            "rootfs directory targets a different CPU architecture \
-             (ELF e_machine {found}, host expects {host})"
-        );
-    }
-
-    let key = canonical_key_for_path(rootfs_dir)
-        .with_context(|| format!("canonicalize rootfs dir {}", rootfs_dir.display()))?;
-    let tree_size = mvm_build::run_image::unpacked_tree_size(rootfs_dir)
-        .with_context(|| format!("measure rootfs dir {}", rootfs_dir.display()))?;
-    let rootfs_rel = format!("rootfs/{key}/rootfs.ext4");
-    let rootfs_abs = cache_root.join(&rootfs_rel);
-    materialize_run_rootfs(&MaterializeExt4Input::new(
-        rootfs_dir.to_path_buf(),
-        rootfs_abs.clone(),
-        tree_size,
-    ))
-    .context("materialize rootfs.ext4 from directory")?;
-
-    let provenance = OciProvenance {
-        schema_version: 1,
-        source: "rootfs_dir".to_string(),
-        supplied_reference: supplied_reference.to_string(),
-        canonical_reference: rootfs_dir.display().to_string(),
-        registry: String::new(),
-        repository: String::new(),
-        tag: None,
-        resolved_digest: String::new(),
-        layer_digests: Vec::new(),
-        trust_policy: "rootfs-dir-unverified".to_string(),
-        verification_status: "no-provenance (dev)".to_string(),
-    };
-    Ok(ResolvedOciRunImage {
-        reference: supplied_reference.to_string(),
-        resolved_digest: String::new(),
-        rootfs_path: rootfs_abs,
-        // rootfs-dir is a niche dev path (no digest); boot it as block ext4.
-        unpacked_root: None,
-        pulled: true,
-        provenance,
-        auth_source: None,
-    })
-}
-
-/// ELF `e_machine` for the host CPU, or `None` for an architecture we don't
-/// map (in which case the rootfs-dir arch guard is skipped rather than
-/// guessing). `EM_X86_64` = 62, `EM_AARCH64` = 183.
-fn host_elf_machine() -> Option<u16> {
-    if cfg!(target_arch = "x86_64") {
-        Some(62)
-    } else if cfg!(target_arch = "aarch64") {
-        Some(183)
-    } else {
-        None
-    }
-}
-
-/// Probe a few representative binaries in `rootfs_dir` and return the first
-/// one's ELF `e_machine`. `None` when none is present/readable as ELF (a tree
-/// with no probe-able binary can't be arch-checked — dev-only, so we don't
-/// block on it).
-fn probe_rootfs_elf_machine(rootfs_dir: &Path) -> Option<u16> {
-    const CANDIDATES: &[&str] = &[
-        "bin/sh",
-        "bin/busybox",
-        "bin/bash",
-        "usr/bin/env",
-        "sbin/init",
-    ];
-    for candidate in CANDIDATES {
-        let path = rootfs_dir.join(candidate);
-        if let Ok(bytes) = read_elf_header(&path)
-            && let Some(machine) = elf_machine(&bytes)
-        {
-            return Some(machine);
-        }
-    }
-    None
-}
-
-/// Read just enough of a file to inspect its ELF header (cheap; never slurps a
-/// whole binary).
-fn read_elf_header(path: &Path) -> std::io::Result<Vec<u8>> {
-    let mut buf = [0u8; 20];
-    let mut file = fs::File::open(path)?;
-    let n = file.read(&mut buf)?;
-    Ok(buf[..n].to_vec())
-}
-
-/// Parse the ELF `e_machine` (offset 18, endianness per `EI_DATA`) from a
-/// header. `None` if the bytes aren't an ELF header.
-fn elf_machine(bytes: &[u8]) -> Option<u16> {
-    if bytes.len() < 20 || bytes[..4] != [0x7f, b'E', b'L', b'F'] {
-        return None;
-    }
-    let m = [bytes[18], bytes[19]];
-    match bytes[5] {
-        1 => Some(u16::from_le_bytes(m)),
-        2 => Some(u16::from_be_bytes(m)),
-        _ => None,
-    }
-}
-
-pub(super) fn pull_image_with_trust(
-    cache_root: &Path,
-    reference: &str,
-    prod: bool,
-) -> Result<(CachedOciImage, OciTrustDecision, String)> {
-    let image_ref: ImageReference = reference.parse()?;
-    if prod && !image_ref.is_digest_pinned() {
-        bail!("mvmctl image pull --prod requires a digest-pinned reference");
-    }
-    pull_image_ref(cache_root, image_ref, reference, prod)
-}
-
-fn trust_decision_for_cached_image(
-    image_ref: &ImageReference,
-    image: &CachedOciImage,
-    prod: bool,
-    verifier: &dyn CosignVerifier,
-) -> Result<OciTrustDecision> {
-    enforce_oci_trust_policy(image_ref, &image.resolved_digest, prod, verifier)
-}
-
-fn enforce_oci_trust_policy(
-    image_ref: &ImageReference,
-    resolved_digest: &str,
-    prod: bool,
-    verifier: &dyn CosignVerifier,
-) -> Result<OciTrustDecision> {
-    if !prod {
-        return Ok(OciTrustDecision::dev_digest_only(image_ref));
-    }
-    let policy = load_oci_registry_policy()?;
-    enforce_oci_trust_policy_with(image_ref, resolved_digest, &policy, verifier)
-}
-
-fn enforce_oci_trust_policy_with(
-    image_ref: &ImageReference,
-    resolved_digest: &str,
-    policy: &OciRegistryPolicy,
-    verifier: &dyn CosignVerifier,
-) -> Result<OciTrustDecision> {
-    enforce_registry_allowlist(image_ref, policy)?;
-    ensure_signature_policy_is_configured(policy)?;
-    let verification_ref = cosign_verification_reference(image_ref, resolved_digest);
-    let mut failures = Vec::new();
-    for identity in &policy.cosign {
-        match verifier.verify(&verification_ref, identity) {
-            Ok(()) => return Ok(OciTrustDecision::cosign_verified(identity)),
-            Err(err) => failures.push(err.to_string()),
-        }
-    }
-    bail!(
-        "cosign verification failed for {} under production OCI policy: {}",
-        verification_ref,
-        failures.join("; ")
-    );
-}
-
-fn enforce_registry_allowlist(
-    image_ref: &ImageReference,
-    policy: &OciRegistryPolicy,
-) -> Result<()> {
-    if !policy.allowed_registries.is_empty()
-        && !policy
-            .allowed_registries
-            .iter()
-            .any(|registry| registry == &image_ref.registry)
-    {
-        bail!(
-            "OCI registry '{}' is denied by production policy",
-            image_ref.registry
-        );
-    }
-    Ok(())
-}
-
-fn ensure_signature_policy_is_configured(policy: &OciRegistryPolicy) -> Result<()> {
-    if !policy.require_signatures {
-        bail!("production OCI policy cannot disable cosign signatures");
-    }
-    if policy.cosign.is_empty() {
-        bail!("production OCI policy requires signatures but has no [[cosign]] trusted identity");
-    }
-    Ok(())
-}
-
-fn cosign_verification_reference(image_ref: &ImageReference, resolved_digest: &str) -> String {
-    format!(
-        "{}/{}@{}",
-        image_ref.registry, image_ref.repository, resolved_digest
-    )
-}
-
-fn load_oci_registry_policy() -> Result<OciRegistryPolicy> {
-    let path = match std::env::var_os("MVM_OCI_POLICY") {
-        Some(path) => PathBuf::from(path),
-        None => PathBuf::from(mvm_core::config::mvm_home()).join("oci-policy.toml"),
-    };
-    if !path.exists() {
-        bail!(
-            "mvmctl image --prod requires an OCI registry policy at {} \
-             (or set MVM_OCI_POLICY to a policy file)",
-            path.display()
-        );
-    }
-    let text = fs::read_to_string(&path)
-        .with_context(|| format!("reading OCI registry policy {}", path.display()))?;
-    parse_oci_registry_policy(&text)
-        .with_context(|| format!("parsing OCI registry policy {}", path.display()))
-}
-
-fn parse_oci_registry_policy(text: &str) -> Result<OciRegistryPolicy> {
-    let policy: OciRegistryPolicy = toml::from_str(text)?;
-    validate_oci_registry_policy(&policy)?;
-    Ok(policy)
-}
-
-fn validate_oci_registry_policy(policy: &OciRegistryPolicy) -> Result<()> {
-    ensure_signature_policy_is_configured(policy)?;
-    for registry in &policy.allowed_registries {
-        if registry.is_empty()
-            || registry.contains("://")
-            || registry.contains('/')
-            || registry.chars().any(char::is_whitespace)
-        {
-            bail!("invalid OCI policy registry host {registry:?}");
-        }
-    }
-    for identity in &policy.cosign {
-        if identity.certificate_identity.is_empty()
-            || identity.certificate_oidc_issuer.is_empty()
-            || identity.certificate_identity.chars().any(char::is_control)
-            || identity
-                .certificate_oidc_issuer
-                .chars()
-                .any(char::is_control)
-        {
-            bail!("invalid empty or control-character cosign identity in OCI policy");
-        }
-    }
-    Ok(())
-}
-
-fn pull_image_ref(
-    cache_root: &Path,
-    image_ref: ImageReference,
-    supplied_reference: &str,
-    prod: bool,
-) -> Result<(CachedOciImage, OciTrustDecision, String)> {
-    let prod_policy = if prod {
-        let policy = load_oci_registry_policy()?;
-        enforce_registry_allowlist(&image_ref, &policy)?;
-        ensure_signature_policy_is_configured(&policy)?;
-        Some(policy)
-    } else {
-        None
-    };
-    let registry_auth = registry_auth_for(&image_ref)?;
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("build Tokio runtime for OCI pull")?;
-
-    let manifest_fetcher = OciManifestFetcher::with_auth(registry_auth.auth);
-    let manifest = runtime
-        .block_on(
-            manifest_fetcher
-                .fetch_linux_platform_manifest(&image_ref, &LinuxPlatform::for_current_arch()),
-        )
-        .context("fetch OCI image manifest")?;
-    let layers = manifest.layers().context("parse OCI image layers")?;
-    if layers.is_empty() {
-        bail!(
-            "OCI image manifest has no layers: {}",
-            image_ref.canonical()
-        );
-    }
-
-    let trust = match &prod_policy {
-        Some(policy) => enforce_oci_trust_policy_with(
-            &image_ref,
-            &manifest.digest,
-            policy,
-            &CosignCommandVerifier,
-        )?,
-        None => OciTrustDecision::dev_digest_only(&image_ref),
-    };
-
-    let manifest_hex = sha256_hex(&manifest.digest)?;
-    let manifest_path = format!("manifests/{manifest_hex}.json");
-    write_cache_file(cache_root, &manifest_path, &manifest.bytes)?;
-
-    let config_path = write_config_blob(
-        cache_root,
-        &runtime,
-        &manifest_fetcher,
-        &image_ref,
-        &manifest.bytes,
-    )?;
-    let layer_fetcher =
-        OciLayerFetcher::from_manifest_fetcher(&manifest_fetcher, LayerFetchOptions::default());
-    let unpacked_root = cache_root.join("unpacked").join(&manifest_hex);
-    if unpacked_root.exists() {
-        fs::remove_dir_all(&unpacked_root)
-            .with_context(|| format!("remove stale unpacked root {}", unpacked_root.display()))?;
-    }
-    fs::create_dir_all(&unpacked_root)
-        .with_context(|| format!("create {}", unpacked_root.display()))?;
-
-    let mut cached_layers = Vec::with_capacity(layers.len());
-    for layer in &layers {
-        let compressed =
-            fetch_or_read_layer(cache_root, &runtime, &layer_fetcher, &image_ref, layer)
-                .with_context(|| format!("fetch layer {}", layer.digest))?;
-        unpack_layer_bytes(layer, &compressed, &unpacked_root)
-            .with_context(|| format!("unpack layer {}", layer.digest))?;
-        cached_layers.push(CachedOciLayer {
-            digest: layer.digest.clone(),
-            size_bytes: layer.size,
-            path: Some(layer_blob_path(&layer.digest)?),
-        });
-    }
-
-    let runtime_tag = oci_runtime_tag();
-    let rootfs_path = format!("rootfs/{manifest_hex}-{runtime_tag}/rootfs.ext4");
-    let rootfs_abs = cache_root.join(&rootfs_path);
-    inject_runtime_and_materialize(
-        cache_root,
-        &unpacked_root,
-        &rootfs_abs,
-        &image_ref.canonical(),
-        None,
-        false,
-    )?;
-
-    let provenance = OciProvenance {
-        schema_version: 1,
-        source: "image_pull".to_string(),
-        supplied_reference: supplied_reference.to_string(),
-        canonical_reference: image_ref.canonical(),
-        registry: image_ref.registry.clone(),
-        repository: image_ref.repository.clone(),
-        tag: image_ref.tag.clone(),
-        resolved_digest: manifest.digest.clone(),
-        layer_digests: cached_layers
-            .iter()
-            .map(|layer| layer.digest.clone())
-            .collect(),
-        trust_policy: trust.trust_policy.clone(),
-        verification_status: trust.verification_status.clone(),
-    };
-    let claims_path = format!("claims/{}.provenance.json", manifest_hex);
-    write_cache_file(
-        cache_root,
-        &claims_path,
-        &serde_json::to_vec_pretty(&provenance).context("serialize OCI provenance")?,
-    )?;
-
-    let mut index = load_index(cache_root)?;
-    let cached = CachedOciImage {
-        reference: image_ref.canonical(),
-        registry: image_ref.registry.clone(),
-        repository: image_ref.repository.clone(),
-        tag: image_ref.tag.clone(),
-        resolved_digest: manifest.digest.clone(),
-        fetched_at: chrono::Utc::now().to_rfc3339_opts(chrono::SecondsFormat::Secs, true),
-        manifest_path,
-        config_path,
-        rootfs_path: Some(rootfs_path),
-        runtime_tag: Some(runtime_tag),
-        claims_path: Some(claims_path),
-        layers: cached_layers,
-    };
-    upsert_image(&mut index, cached.clone());
-    save_index(cache_root, &index)?;
-    Ok((cached, trust, registry_auth.source))
-}
-
-fn write_config_blob(
-    cache_root: &Path,
-    runtime: &tokio::runtime::Runtime,
-    manifest_fetcher: &OciManifestFetcher,
-    image_ref: &ImageReference,
-    manifest_bytes: &[u8],
-) -> Result<Option<String>> {
-    let Some(config) = manifest_config_descriptor(manifest_bytes)? else {
-        return Ok(None);
-    };
-    let config_path = format!("configs/{}.json", sha256_hex(&config.digest)?);
-    if let Some(bytes) = read_verified_cache_file(cache_root, &config_path, &config.digest)?
-        && serde_json::from_slice::<Value>(&bytes).is_ok()
-    {
-        return Ok(Some(config_path));
-    }
-
-    let fetcher =
-        OciLayerFetcher::from_manifest_fetcher(manifest_fetcher, LayerFetchOptions::default());
-    let mut bytes = Vec::new();
-    runtime
-        .block_on(fetcher.fetch_layer(image_ref, &config, &mut bytes))
-        .context("fetch OCI image config blob")?;
-    write_cache_file(cache_root, &config_path, &bytes)?;
-    Ok(Some(config_path))
-}
-
-fn manifest_config_descriptor(manifest_bytes: &[u8]) -> Result<Option<LayerDescriptor>> {
-    let value: Value = serde_json::from_slice(manifest_bytes).context("parse manifest JSON")?;
-    let Some(config) = value.get("config").and_then(Value::as_object) else {
-        return Ok(None);
-    };
-    let digest = config
-        .get("digest")
-        .and_then(Value::as_str)
-        .context("manifest config missing digest")?
-        .to_string();
-    let media_type = config
-        .get("mediaType")
-        .and_then(Value::as_str)
-        .unwrap_or("application/vnd.oci.image.config.v1+json")
-        .to_string();
-    let size = config.get("size").and_then(Value::as_u64).unwrap_or(0);
-    Ok(Some(LayerDescriptor {
-        digest,
-        size,
-        media_type,
-    }))
-}
-
-fn fetch_or_read_layer(
-    cache_root: &Path,
-    runtime: &tokio::runtime::Runtime,
-    fetcher: &OciLayerFetcher,
-    image_ref: &ImageReference,
-    layer: &LayerDescriptor,
-) -> Result<Vec<u8>> {
-    let path = layer_blob_path(&layer.digest)?;
-    if let Some(bytes) = read_verified_cache_file(cache_root, &path, &layer.digest)? {
-        return Ok(bytes);
-    }
-    let mut bytes = Vec::new();
-    runtime.block_on(fetcher.fetch_layer(image_ref, layer, &mut bytes))?;
-    write_cache_file(cache_root, &path, &bytes)?;
-    Ok(bytes)
-}
-
-fn unpack_layer_bytes(layer: &LayerDescriptor, bytes: &[u8], unpacked_root: &Path) -> Result<()> {
-    let report = if is_gzip_layer(&layer.media_type) {
-        unpack_layer(
-            GzDecoder::new(Cursor::new(bytes)),
-            unpacked_root,
-            &UnpackOptions::default(),
-        )
-    } else {
-        unpack_layer(Cursor::new(bytes), unpacked_root, &UnpackOptions::default())
-    }?;
-    if !report.refused.is_empty() {
-        bail!("layer unpack refused entries: {:?}", report.refused);
-    }
-    Ok(())
-}
-
-fn is_gzip_layer(media_type: &str) -> bool {
-    media_type.ends_with("+gzip")
-        || media_type.ends_with(".gzip")
-        || media_type.contains("tar.gzip")
-}
-
-pub(super) fn list_rows(cache_root: &Path, registry: Option<&str>) -> Result<Vec<ImageListRow>> {
-    let index = load_index(cache_root)?;
-    let rows = index
-        .images
-        .iter()
-        .filter(|image| registry.is_none_or(|want| image.registry == want))
-        .map(|image| ImageListRow {
-            reference: image.reference.clone(),
-            registry: image.registry.clone(),
-            repository: image.repository.clone(),
-            tag: image.tag.clone(),
-            resolved_digest: image.resolved_digest.clone(),
-            fetched_at: image.fetched_at.clone(),
-            size_bytes: image_size_bytes(cache_root, image),
-            layers: image.layers.len(),
-        })
-        .collect();
-    Ok(rows)
-}
-
-pub(super) fn inspect_image(cache_root: &Path, reference: &str) -> Result<InspectOutput> {
-    let index = load_index(cache_root)?;
-    let image = find_image(&index, reference)
-        .with_context(|| format!("cached OCI image not found for '{reference}'"))?
-        .clone();
-    Ok(InspectOutput {
-        size_bytes: image_size_bytes(cache_root, &image),
-        manifest: read_json_optional(cache_root, &image.manifest_path)?,
-        config: image
-            .config_path
-            .as_deref()
-            .map(|p| read_json_optional(cache_root, p))
-            .transpose()?
-            .flatten(),
-        claims: image
-            .claims_path
-            .as_deref()
-            .map(|p| read_json_optional(cache_root, p))
-            .transpose()?
-            .flatten(),
-        image,
-    })
-}
-
-pub(super) fn remove_image(cache_root: &Path, reference: &str) -> Result<RemoveOutcome> {
-    let mut index = load_index(cache_root)?;
-    let Some(position) = index
-        .images
-        .iter()
-        .position(|image| image_matches(image, reference))
-    else {
-        bail!("cached OCI image not found for '{reference}'");
-    };
-
-    let image = index.images.remove(position);
-    let mut removed_files = 0usize;
-    let mut freed_bytes = 0u64;
-    let shared_layer_paths = remaining_layer_paths(&index);
-    validate_image_paths(cache_root, &image)?;
-
-    for path in metadata_paths(&image) {
-        remove_cache_file(cache_root, &path, &mut removed_files, &mut freed_bytes)?;
-    }
-
-    for layer in &image.layers {
-        let Some(path) = layer.path.as_deref() else {
-            continue;
-        };
-        if shared_layer_paths.contains(path) {
-            continue;
-        }
-        remove_cache_file(cache_root, path, &mut removed_files, &mut freed_bytes)?;
-    }
-
-    save_index(cache_root, &index)?;
-    Ok(RemoveOutcome {
-        reference: image.reference,
-        removed_files,
-        freed_bytes,
-    })
-}
-
-fn load_index(cache_root: &Path) -> Result<OciCacheIndex> {
-    let path = cache_root.join(INDEX_FILE);
-    if !path.exists() {
-        return Ok(OciCacheIndex::default());
-    }
-    let bytes = fs::read(&path).with_context(|| format!("read {}", path.display()))?;
-    let index: OciCacheIndex =
-        serde_json::from_slice(&bytes).with_context(|| format!("parse {}", path.display()))?;
-    if index.schema_version != 1 {
-        bail!(
-            "unsupported OCI cache index schema_version {} at {}",
-            index.schema_version,
-            path.display()
-        );
-    }
-    Ok(index)
-}
-
-fn save_index(cache_root: &Path, index: &OciCacheIndex) -> Result<()> {
-    fs::create_dir_all(cache_root).with_context(|| format!("create {}", cache_root.display()))?;
-    let path = cache_root.join(INDEX_FILE);
-    let bytes = serde_json::to_vec_pretty(index).context("serialize OCI cache index")?;
-    fs::write(&path, bytes).with_context(|| format!("write {}", path.display()))
-}
-
-fn read_verified_cache_file(
-    cache_root: &Path,
-    relative: &str,
-    digest: &str,
-) -> Result<Option<Vec<u8>>> {
-    let path = safe_cache_path(cache_root, relative)?;
-    if !path.exists() {
-        return Ok(None);
-    }
-    let bytes = fs::read(&path).with_context(|| format!("read {}", path.display()))?;
-    verify_sha256_digest(&bytes, digest)
-        .with_context(|| format!("verify cached blob {}", path.display()))?;
-    Ok(Some(bytes))
-}
-
-fn write_cache_file(cache_root: &Path, relative: &str, bytes: &[u8]) -> Result<()> {
-    let path = safe_cache_path(cache_root, relative)?;
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent).with_context(|| format!("create {}", parent.display()))?;
-    }
-    fs::write(&path, bytes).with_context(|| format!("write {}", path.display()))
-}
-
-fn layer_blob_path(digest: &str) -> Result<String> {
-    Ok(format!("blobs/sha256/{}", sha256_hex(digest)?))
-}
-
-/// The unpacked+injected OCI tree for `resolved_digest`, iff present on disk.
-/// Its location is deterministic (`<cache>/unpacked/<sha256_hex(digest)>`); a
-/// virtiofs-root dev boot serves it directly, so gate on existence.
-fn unpacked_dir_if_present(cache_root: &Path, resolved_digest: &str) -> Option<PathBuf> {
-    if resolved_digest.is_empty() {
-        return None;
-    }
-    let hex = sha256_hex(resolved_digest).ok()?;
-    let dir = cache_root.join("unpacked").join(hex);
-    dir.is_dir().then_some(dir)
-}
-
-fn sha256_hex(digest: &str) -> Result<String> {
-    let Some(hex) = digest.strip_prefix("sha256:") else {
-        bail!("unsupported digest algorithm in {digest:?}");
-    };
-    if hex.len() != 64
-        || !hex
-            .bytes()
-            .all(|b| b.is_ascii_hexdigit() && !b.is_ascii_uppercase())
-    {
-        bail!("malformed sha256 digest: {digest:?}");
-    }
-    Ok(hex.to_string())
-}
-
-fn upsert_image(index: &mut OciCacheIndex, image: CachedOciImage) {
-    if let Some(existing) = index
-        .images
-        .iter_mut()
-        .find(|cached| cached.reference == image.reference)
-    {
-        *existing = image;
-    } else {
-        index.images.push(image);
-    }
-}
-
-fn find_image<'a>(index: &'a OciCacheIndex, reference: &str) -> Option<&'a CachedOciImage> {
-    index
-        .images
-        .iter()
-        .find(|image| image_matches(image, reference))
-}
-
-fn image_matches(image: &CachedOciImage, reference: &str) -> bool {
-    image.reference == reference
-        || image.resolved_digest == reference
-        || image
-            .resolved_digest
-            .strip_prefix("sha256:")
-            .is_some_and(|digest| digest == reference)
-}
-
-fn image_size_bytes(cache_root: &Path, image: &CachedOciImage) -> u64 {
-    let mut total = 0u64;
-    let mut seen = BTreeSet::new();
-    for path in all_image_paths(image) {
-        if !seen.insert(path.clone()) {
-            continue;
-        }
-        if let Ok(path) = safe_cache_path(cache_root, &path)
-            && let Ok(meta) = path.metadata()
-            && meta.is_file()
-        {
-            total = total.saturating_add(meta.len());
-        }
-    }
-    if total == 0 {
-        image
-            .layers
-            .iter()
-            .map(|layer| layer.size_bytes)
-            .fold(0u64, u64::saturating_add)
-    } else {
-        total
-    }
-}
-
-fn all_image_paths(image: &CachedOciImage) -> Vec<String> {
-    let mut paths = metadata_paths(image);
-    paths.extend(image.layers.iter().filter_map(|layer| layer.path.clone()));
-    paths
-}
-
-fn metadata_paths(image: &CachedOciImage) -> Vec<String> {
-    let mut paths = vec![image.manifest_path.clone()];
-    paths.extend(image.config_path.clone());
-    paths.extend(image.rootfs_path.clone());
-    paths.extend(image.claims_path.clone());
-    paths
-}
-
-fn remaining_layer_paths(index: &OciCacheIndex) -> BTreeSet<String> {
-    index
-        .images
-        .iter()
-        .flat_map(|image| image.layers.iter())
-        .filter_map(|layer| layer.path.clone())
-        .collect()
-}
-
-fn validate_image_paths(cache_root: &Path, image: &CachedOciImage) -> Result<()> {
-    for path in all_image_paths(image) {
-        let _ = safe_cache_path(cache_root, &path)?;
-    }
-    Ok(())
-}
-
-fn read_json_optional(cache_root: &Path, relative: &str) -> Result<Option<Value>> {
-    let path = safe_cache_path(cache_root, relative)?;
-    if !path.exists() {
-        return Ok(None);
-    }
-    let bytes = fs::read(&path).with_context(|| format!("read {}", path.display()))?;
-    serde_json::from_slice(&bytes)
-        .map(Some)
-        .with_context(|| format!("parse {}", path.display()))
-}
-
-fn remove_cache_file(
-    cache_root: &Path,
-    relative: &str,
-    removed_files: &mut usize,
-    freed_bytes: &mut u64,
-) -> Result<()> {
-    let path = safe_cache_path(cache_root, relative)?;
-    if !path.exists() {
-        return Ok(());
-    }
-    let meta = path
-        .metadata()
-        .with_context(|| format!("stat {}", path.display()))?;
-    if !meta.is_file() {
-        bail!("refusing to remove non-file cache path {}", path.display());
-    }
-    let len = meta.len();
-    fs::remove_file(&path).with_context(|| format!("remove {}", path.display()))?;
-    *removed_files += 1;
-    *freed_bytes = freed_bytes.saturating_add(len);
-    prune_empty_parents(cache_root, path.parent())?;
-    Ok(())
-}
-
-fn prune_empty_parents(cache_root: &Path, mut current: Option<&Path>) -> Result<()> {
-    while let Some(dir) = current {
-        if dir == cache_root {
-            break;
-        }
-        match fs::remove_dir(dir) {
-            Ok(()) => current = dir.parent(),
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => current = dir.parent(),
-            Err(e) if e.kind() == std::io::ErrorKind::DirectoryNotEmpty => break,
-            Err(e) => return Err(e).with_context(|| format!("remove {}", dir.display())),
-        }
-    }
-    Ok(())
-}
-
-fn safe_cache_path(cache_root: &Path, relative: &str) -> Result<PathBuf> {
-    let rel = Path::new(relative);
-    if rel.is_absolute() {
-        bail!("OCI cache path must be relative: {relative}");
-    }
-    if rel.components().any(|component| {
-        matches!(
-            component,
-            Component::ParentDir | Component::RootDir | Component::Prefix(_)
-        )
-    }) {
-        bail!("OCI cache path escapes cache root: {relative}");
-    }
-    Ok(cache_root.join(rel))
-}
-
-pub(super) fn render_list(rows: &[ImageListRow]) {
-    if rows.is_empty() {
-        println!("No cached OCI images.");
-        return;
-    }
-    println!(
-        "{:<38} {:<20} {:<18} {:>10}",
-        "REFERENCE", "DIGEST", "FETCHED", "SIZE"
-    );
-    for row in rows {
-        println!(
-            "{:<38} {:<20} {:<18} {:>10}",
-            truncate(&row.reference, 38),
-            truncate(&row.resolved_digest, 20),
-            truncate(&row.fetched_at, 18),
-            human_bytes(row.size_bytes)
-        );
-    }
-}
-
-pub(super) fn render_inspect(output: &InspectOutput) {
-    let image = &output.image;
-    println!("Reference: {}", image.reference);
-    println!("Registry: {}", image.registry);
-    println!("Repository: {}", image.repository);
-    if let Some(tag) = &image.tag {
-        println!("Tag: {tag}");
-    }
-    println!("Resolved digest: {}", image.resolved_digest);
-    println!("Fetched at: {}", image.fetched_at);
-    println!("Size: {}", human_bytes(output.size_bytes));
-    println!("Manifest: {}", image.manifest_path);
-    if let Some(path) = &image.config_path {
-        println!("Config: {path}");
-    }
-    if let Some(path) = &image.rootfs_path {
-        println!("Rootfs: {path}");
-    }
-    println!(
-        "mvm-claims.json: {}",
-        if output.claims.is_some() {
-            "present"
-        } else {
-            "absent"
-        }
-    );
-    println!("Layers:");
-    for layer in &image.layers {
-        let path = layer.path.as_deref().unwrap_or("-");
-        println!(
-            "  {}  {}  {}",
-            layer.digest,
-            human_bytes(layer.size_bytes),
-            path
-        );
-    }
-    if let Some(claims) = &output.claims
-        && let Some(labels) = claims.as_object().and_then(|obj| obj.get("labels"))
-    {
-        let labels: BTreeMap<_, _> = labels
-            .as_object()
-            .into_iter()
-            .flat_map(|obj| obj.iter())
-            .collect();
-        if !labels.is_empty() {
-            println!("Claim labels:");
-            for (key, value) in labels {
-                println!("  {key}: {value}");
-            }
-        }
-    }
-}
-
-fn truncate(value: &str, max: usize) -> String {
-    if value.chars().count() <= max {
-        return value.to_string();
-    }
-    let mut s: String = value.chars().take(max.saturating_sub(1)).collect();
-    s.push('~');
-    s
-}
-
-#[cfg(test)]
-mod tests {
-    use super::trust::CosignVerifyError;
-    use super::*;
-    use flate2::{Compression, write::GzEncoder};
-    use sha2::{Digest, Sha256};
-    use std::cell::RefCell;
-    use tar::{Builder, EntryType, Header};
-
-    // The default run-path materialize is the pure in-process writer, and it
-    // must emit dm-verity sidecars so OCI block roots boot sealed even before
-    // the builder-VM fallback is exercised.
-    #[cfg(feature = "pure-mkfs")]
-    #[test]
-    fn materialize_run_rootfs_default_is_pure_and_verity_backed() {
-        // The builder-VM escape hatch must be unset for the default (pure) path.
-        assert!(
-            std::env::var_os("MVM_MATERIALIZE_BUILDER_VM").is_none(),
-            "test env must not force the builder-VM materializer"
-        );
-        let src = tempfile::tempdir().unwrap();
-        std::fs::create_dir(src.path().join("etc")).unwrap();
-        std::fs::write(src.path().join("etc/hostname"), b"box\n").unwrap();
-        let out = tempfile::tempdir().unwrap();
-        let rootfs = out.path().join("rootfs.ext4");
-
-        materialize_run_rootfs(&MaterializeExt4Input::new(
-            src.path().to_path_buf(),
-            rootfs.clone(),
-            0,
-        ))
-        .expect("pure materialize");
-
-        // A real ext4 image was written in-process (superblock magic present)…
-        let img = std::fs::read(&rootfs).unwrap();
-        assert_eq!(&img[1024 + 0x38..1024 + 0x3A], &[0x53, 0xEF]);
-        // …and the sealed-boot sidecars sit beside it.
-        assert!(out.path().join("rootfs.verity").exists());
-        let roothash = std::fs::read_to_string(out.path().join("rootfs.roothash")).unwrap();
-        assert_eq!(roothash.trim().len(), 64);
-    }
-
-    struct MockCosignVerifier {
-        results: RefCell<Vec<Result<(), CosignVerifyError>>>,
-    }
-
-    impl MockCosignVerifier {
-        fn new(results: Vec<Result<(), CosignVerifyError>>) -> Self {
-            Self {
-                results: RefCell::new(results),
-            }
-        }
-    }
-
-    impl CosignVerifier for MockCosignVerifier {
-        fn verify(
-            &self,
-            _reference: &str,
-            _identity: &CosignIdentity,
-        ) -> Result<(), CosignVerifyError> {
-            self.results.borrow_mut().remove(0)
-        }
-    }
-
-    fn sample_image(reference: &str, digest: &str, layer_path: &str) -> CachedOciImage {
-        CachedOciImage {
-            reference: reference.to_string(),
-            registry: "docker.io".to_string(),
-            repository: "library/alpine".to_string(),
-            tag: Some("3.20".to_string()),
-            resolved_digest: digest.to_string(),
-            fetched_at: "2026-05-18T00:00:00Z".to_string(),
-            manifest_path: "manifests/alpine.json".to_string(),
-            config_path: Some("configs/alpine.json".to_string()),
-            rootfs_path: None,
-            runtime_tag: None,
-            claims_path: Some("claims/alpine.json".to_string()),
-            layers: vec![CachedOciLayer {
-                digest: "sha256:layer".to_string(),
-                size_bytes: 4,
-                path: Some(layer_path.to_string()),
-            }],
-        }
-    }
-
-    fn write_index(cache_root: &Path, index: &OciCacheIndex) {
-        fs::create_dir_all(cache_root).expect("create cache root");
-        fs::write(
-            cache_root.join(INDEX_FILE),
-            serde_json::to_vec_pretty(index).expect("serialize index"),
-        )
-        .expect("write index");
-    }
-
-    #[test]
-    fn default_index_round_trips_through_save_and_load() {
-        // Regression: a derived `Default` set schema_version = 0, which
-        // save_index persisted and load_index then rejected — breaking
-        // `image ls` / `run --image` on a freshly-created OCI cache.
-        let tmp = tempfile::tempdir().expect("tempdir");
-        assert_eq!(OciCacheIndex::default().schema_version, schema_version());
-        save_index(tmp.path(), &OciCacheIndex::default()).expect("save a default index");
-        let loaded = load_index(tmp.path()).expect("a freshly-saved default index must load");
-        assert_eq!(loaded.schema_version, schema_version());
-        assert!(loaded.images.is_empty());
-    }
-
-    fn write_file(cache_root: &Path, relative: &str, body: &[u8]) {
-        let path = cache_root.join(relative);
-        fs::create_dir_all(path.parent().expect("relative has parent")).expect("create parent");
-        fs::write(path, body).expect("write cache file");
-    }
-
-    fn write_minimal_config(cache_root: &Path) {
-        write_file(cache_root, "configs/alpine.json", br#"{"config":{}}"#);
-    }
-
-    fn create_unpacked_root(cache_root: &Path, digest: &str) -> PathBuf {
-        let unpacked = cache_root
-            .join("unpacked")
-            .join(sha256_hex(digest).unwrap());
-        fs::create_dir_all(&unpacked).expect("create unpacked root");
-        fs::write(unpacked.join("layer-file"), b"from-layer").expect("write unpacked file");
-        unpacked
-    }
-
-    fn seed_guest_runtime_cache(cache_root: &Path) {
-        use mvm_build::guest_agent_build::GuestAgentLayout;
-        use mvm_core::arch::GuestArch;
-
-        let guest_layout =
-            GuestAgentLayout::under(cache_root, env!("CARGO_PKG_VERSION"), GuestArch::host());
-        std::fs::create_dir_all(&guest_layout.dir).expect("create guest cache dir");
-        for path in [
-            &guest_layout.oci_init,
-            &guest_layout.agent,
-            &guest_layout.netinit,
-            &guest_layout.netd,
-            &guest_layout.egress_client,
-            &guest_layout.entrypoint_runner,
-            &guest_layout.verity_init,
-        ] {
-            std::fs::write(path, b"#!/bin/sh\nexit 0\n").expect("seed guest runtime cache");
-        }
-    }
-
-    fn fake_runtime_materialize(
-        _cache_root: &Path,
-        unpacked_root: &Path,
-        rootfs_abs: &Path,
-        image_label: &str,
-        _entrypoint: Option<&OciEntrypointConfig>,
-        _sealed: bool,
-    ) -> Result<()> {
-        assert!(unpacked_root.is_dir(), "unpacked root must exist");
-        let parent = rootfs_abs.parent().expect("rootfs has parent");
-        fs::create_dir_all(parent)?;
-        fs::write(rootfs_abs, format!("materialized:{image_label}"))?;
-        fs::write(parent.join("rootfs.verity"), b"fake-verity")?;
-        fs::write(parent.join("rootfs.roothash"), b"abc\n")?;
-        Ok(())
-    }
-
-    fn digest_of(bytes: &[u8]) -> String {
-        format!("sha256:{}", hex::encode(Sha256::digest(bytes)))
-    }
-
-    fn add_tar_entry(builder: &mut Builder<Vec<u8>>, name: &str, body: &[u8]) {
-        let mut header = Header::new_gnu();
-        header.set_size(body.len() as u64);
-        header.set_entry_type(EntryType::Regular);
-        header.set_mode(0o644);
-        header.set_cksum();
-        builder
-            .append_data(&mut header, name, body)
-            .expect("append tar entry");
-    }
-
-    fn traversal_layer_gzip() -> Vec<u8> {
-        let mut buf = Vec::new();
-        {
-            let gz = GzEncoder::new(&mut buf, Compression::default());
-            let mut tar = Builder::new(gz);
-            tar.mode(tar::HeaderMode::Deterministic);
-
-            let mut bad = Header::new_gnu();
-            bad.set_size(1);
-            bad.set_mode(0o644);
-            bad.set_mtime(0);
-            bad.set_entry_type(EntryType::Regular);
-            let name_bytes = b"../escape";
-            let name_field = bad.as_old_mut().name.as_mut_slice();
-            name_field[..name_bytes.len()].copy_from_slice(name_bytes);
-            bad.set_cksum();
-            tar.append(&bad, &b"x"[..]).expect("append traversal entry");
-            tar.finish().expect("finish layer tar");
-        }
-        buf
-    }
-
-    fn oci_archive_with_layer(layer: &[u8]) -> Vec<u8> {
-        let platform = LinuxPlatform::for_current_arch();
-        let layer_digest = digest_of(layer);
-        let config = format!(
-            r#"{{"architecture":"{}","os":"linux","rootfs":{{"type":"layers","diff_ids":["{}"]}}}}"#,
-            platform.architecture, layer_digest
-        )
-        .into_bytes();
-        let config_digest = digest_of(&config);
-        let manifest = format!(
-            r#"{{"schemaVersion":2,"mediaType":"application/vnd.oci.image.manifest.v1+json","config":{{"mediaType":"application/vnd.oci.image.config.v1+json","digest":"{config_digest}","size":{config_size}}},"layers":[{{"mediaType":"application/vnd.oci.image.layer.v1.tar+gzip","digest":"{layer_digest}","size":{layer_size}}}]}}"#,
-            config_size = config.len(),
-            layer_size = layer.len()
-        )
-        .into_bytes();
-        let manifest_digest = digest_of(&manifest);
-        let index = format!(
-            r#"{{"schemaVersion":2,"mediaType":"application/vnd.oci.image.index.v1+json","manifests":[{{"mediaType":"application/vnd.oci.image.manifest.v1+json","digest":"{manifest_digest}","size":{manifest_size},"platform":{{"architecture":"{}","os":"linux"}}}}]}}"#,
-            platform.architecture,
-            manifest_size = manifest.len()
-        )
-        .into_bytes();
-
-        let mut builder = Builder::new(Vec::new());
-        add_tar_entry(
-            &mut builder,
-            "oci-layout",
-            br#"{"imageLayoutVersion":"1.0.0"}"#,
-        );
-        add_tar_entry(&mut builder, "index.json", &index);
-        add_tar_entry(
-            &mut builder,
-            &format!(
-                "blobs/sha256/{}",
-                manifest_digest
-                    .strip_prefix("sha256:")
-                    .expect("manifest digest prefix")
-            ),
-            &manifest,
-        );
-        add_tar_entry(
-            &mut builder,
-            &format!(
-                "blobs/sha256/{}",
-                config_digest
-                    .strip_prefix("sha256:")
-                    .expect("config digest prefix")
-            ),
-            &config,
-        );
-        add_tar_entry(
-            &mut builder,
-            &format!(
-                "blobs/sha256/{}",
-                layer_digest
-                    .strip_prefix("sha256:")
-                    .expect("layer digest prefix")
-            ),
-            layer,
-        );
-        builder.into_inner().expect("finish OCI archive")
-    }
-
-    #[test]
-    fn prod_pull_requires_digest_pin_before_network() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let err = pull_image_with_trust(tmp.path(), "docker.io/library/alpine:3.20", true)
-            .expect_err("mutable prod pull must fail before registry access");
-        assert!(
-            err.to_string()
-                .contains("requires a digest-pinned reference"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn local_archive_prod_is_refused() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let err = ingest_local_archive(
-            tmp.path(),
-            Path::new("/does/not/matter.tar"),
-            "oci-archive:/does/not/matter.tar",
-            true,
-        )
-        .expect_err("prod local archive must be refused");
-        assert!(err.to_string().contains("prod"), "got {err}");
-    }
-
-    #[test]
-    fn local_archive_missing_file_errors() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let missing = tmp.path().join("nope.tar");
-        let err = ingest_local_archive(tmp.path(), &missing, "oci-archive:nope.tar", false)
-            .expect_err("missing archive must error");
-        assert!(err.to_string().contains("open OCI archive"), "got {err}");
-    }
-
-    #[test]
-    fn local_archive_malformed_is_rejected() {
-        // Fails inside read_oci_archive (no oci-layout / index) before any
-        // materialize, so no builder VM is needed to exercise the reject path.
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let bad = tmp.path().join("bad.tar");
-        std::fs::write(&bad, b"definitely not a valid tar archive").unwrap();
-        let err = ingest_local_archive(tmp.path(), &bad, "oci-archive:bad.tar", false)
-            .expect_err("malformed archive must be rejected");
-        assert!(!err.to_string().is_empty());
-    }
-
-    #[test]
-    fn local_archive_layer_traversal_is_rejected_by_hardened_unpack() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let archive = oci_archive_with_layer(&traversal_layer_gzip());
-        let err = ingest_archive_from_reader(
-            tmp.path(),
-            Cursor::new(archive),
-            "oci-archive:traversal.tar",
-            "local_archive",
-            "OCI archive with traversal layer",
-        )
-        .expect_err("local archive ingest must reject traversal entries");
-        let msg = format!("{err:#}");
-        assert!(
-            msg.contains("refused") || msg.contains("traversal"),
-            "got {msg}"
-        );
-    }
-
-    #[test]
-    fn stdin_archive_prod_is_refused() {
-        // prod bails before any stdin read.
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let err = ingest_stdin_archive(tmp.path(), "-", true)
-            .expect_err("prod stdin archive must be refused");
-        assert!(err.to_string().contains("prod"), "got {err}");
-    }
-
-    fn fake_elf(machine: u16) -> Vec<u8> {
-        let mut b = vec![0u8; 20];
-        b[0..4].copy_from_slice(&[0x7f, b'E', b'L', b'F']);
-        b[4] = 2; // ELFCLASS64
-        b[5] = 1; // little-endian
-        b[6] = 1; // version
-        b[16..18].copy_from_slice(&2u16.to_le_bytes()); // e_type ET_EXEC
-        b[18..20].copy_from_slice(&machine.to_le_bytes());
-        b
-    }
-
-    #[test]
-    fn elf_machine_parses_le_and_rejects_non_elf() {
-        assert_eq!(elf_machine(&fake_elf(62)), Some(62));
-        assert_eq!(elf_machine(&fake_elf(183)), Some(183));
-        assert_eq!(elf_machine(b"definitely not an elf!"), None);
-    }
-
-    #[test]
-    fn rootfs_dir_prod_is_refused() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let err = ingest_rootfs_dir(tmp.path(), tmp.path(), "rootfs-dir:.", true)
-            .expect_err("prod rootfs dir must be refused");
-        assert!(err.to_string().contains("prod"), "got {err}");
-    }
-
-    #[test]
-    fn rootfs_dir_missing_is_rejected() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let missing = tmp.path().join("nope");
-        let err = ingest_rootfs_dir(tmp.path(), &missing, "rootfs-dir:nope", false)
-            .expect_err("missing rootfs dir must error");
-        assert!(err.to_string().contains("directory"), "got {err}");
-    }
-
-    #[test]
-    fn rootfs_dir_wrong_arch_is_rejected() {
-        // The guard is a no-op on hosts we don't map; skip there.
-        let Some(host) = host_elf_machine() else {
-            return;
-        };
-        let wrong = if host == 62 { 183 } else { 62 };
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let bin = tmp.path().join("bin");
-        std::fs::create_dir_all(&bin).unwrap();
-        std::fs::write(bin.join("sh"), fake_elf(wrong)).unwrap();
-        let err = ingest_rootfs_dir(tmp.path(), tmp.path(), "rootfs-dir:.", false)
-            .expect_err("cross-arch rootfs must be rejected before materialize");
-        assert!(err.to_string().contains("architecture"), "got {err}");
-    }
-
-    #[test]
-    fn archive_from_reader_rejects_malformed() {
-        // The shared file/stdin core rejects a garbage stream inside
-        // read_oci_archive, before any materialize (no builder VM needed).
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let err = ingest_archive_from_reader(
-            tmp.path(),
-            Cursor::new(b"not a valid oci archive".to_vec()),
-            "-",
-            "stdin_archive",
-            "OCI archive on stdin",
-        )
-        .expect_err("malformed stdin archive must be rejected");
-        assert!(!err.to_string().is_empty());
-    }
-
-    #[test]
-    fn prod_run_image_requires_digest_pin_before_network() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let err = resolve_or_pull_run_image(tmp.path(), "docker.io/library/alpine:3.20", true)
-            .expect_err("mutable prod run image must fail before registry access");
-        assert!(
-            err.to_string()
-                .contains("requires a digest-pinned reference"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn runtime_tag_includes_guest_runtime_hash() {
-        let tag = oci_runtime_tag();
-        assert!(
-            tag.starts_with(concat!(env!("CARGO_PKG_VERSION"), "-guest-")),
-            "unexpected runtime tag: {tag}"
-        );
-        assert_eq!(
-            tag.rsplit_once("-guest-").unwrap().1.len(),
-            16,
-            "runtime tag should carry a short guest-runtime digest"
-        );
-    }
-
-    #[test]
-    fn resolve_run_image_uses_cached_rootfs() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        let mut image = sample_image("docker.io/library/alpine:3.20", digest, "blobs/a");
-        image.rootfs_path = Some("rootfs/alpine/rootfs.ext4".to_string());
-        image.runtime_tag = Some(oci_runtime_tag());
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-        write_file(tmp.path(), "rootfs/alpine/rootfs.ext4", b"rootfs");
-        write_file(tmp.path(), "rootfs/alpine/rootfs.verity", b"verity");
-        write_file(
-            tmp.path(),
-            "rootfs/alpine/rootfs.roothash",
-            b"0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef\n",
-        );
-
-        let resolved =
-            resolve_or_pull_run_image(tmp.path(), "docker.io/library/alpine:3.20", false)
-                .expect("cached rootfs resolves");
-
-        assert_eq!(resolved.reference, "docker.io/library/alpine:3.20");
-        assert_eq!(resolved.resolved_digest, digest);
-        assert!(resolved.rootfs_path.ends_with("rootfs/alpine/rootfs.ext4"));
-        assert!(!resolved.pulled);
-        assert_eq!(resolved.provenance.source, "run_image");
-        assert_eq!(
-            resolved.provenance.supplied_reference,
-            "docker.io/library/alpine:3.20"
-        );
-        assert_eq!(resolved.provenance.registry, "docker.io");
-        assert_eq!(
-            resolved.provenance.layer_digests,
-            vec!["sha256:layer".to_string()]
-        );
-    }
-
-    #[test]
-    fn resolve_run_image_rematerializes_stale_record_without_rootfs_path() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        let image = sample_image("docker.io/library/alpine:3.20", digest, "blobs/a");
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-        write_minimal_config(tmp.path());
-        create_unpacked_root(tmp.path(), digest);
-        seed_guest_runtime_cache(tmp.path());
-
-        let resolved = resolve_or_pull_run_image_with(
-            tmp.path(),
-            "docker.io/library/alpine:3.20",
-            false,
-            fake_runtime_materialize,
-        )
-        .expect("stale cached image should be repaired from unpacked layers");
-
-        let runtime_tag = oci_runtime_tag();
-        let expected = format!(
-            "rootfs/{}-{runtime_tag}/rootfs.ext4",
-            sha256_hex(digest).unwrap()
-        );
-        assert_eq!(resolved.rootfs_path, tmp.path().join(&expected));
-        assert!(!resolved.pulled);
-        assert_eq!(
-            fs::read_to_string(&resolved.rootfs_path).expect("read repaired rootfs"),
-            "materialized:docker.io/library/alpine:3.20"
-        );
-        let index = load_index(tmp.path()).expect("load repaired index");
-        let repaired = find_image(&index, "docker.io/library/alpine:3.20")
-            .expect("repaired image still indexed");
-        assert_eq!(repaired.rootfs_path.as_deref(), Some(expected.as_str()));
-        assert_eq!(repaired.runtime_tag.as_deref(), Some(runtime_tag.as_str()));
-    }
-
-    #[test]
-    fn resolve_run_image_rematerializes_missing_current_rootfs() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        let mut image = sample_image("docker.io/library/alpine:3.20", digest, "blobs/a");
-        image.rootfs_path = Some("rootfs/alpine/rootfs.ext4".to_string());
-        image.runtime_tag = Some(oci_runtime_tag());
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-        write_minimal_config(tmp.path());
-        create_unpacked_root(tmp.path(), digest);
-        seed_guest_runtime_cache(tmp.path());
-
-        let resolved = resolve_or_pull_run_image_with(
-            tmp.path(),
-            "docker.io/library/alpine:3.20",
-            false,
-            fake_runtime_materialize,
-        )
-        .expect("missing current rootfs should be repaired from unpacked layers");
-
-        assert_eq!(
-            resolved.rootfs_path,
-            tmp.path().join("rootfs/alpine/rootfs.ext4")
-        );
-        assert!(!resolved.pulled);
-        assert_eq!(
-            fs::read_to_string(&resolved.rootfs_path).expect("read repaired rootfs"),
-            "materialized:docker.io/library/alpine:3.20"
-        );
-    }
-
-    #[test]
-    fn resolve_run_image_missing_rootfs_without_unpacked_layers_asks_for_repull() {
-        // The index records a materialized rootfs whose ext4 has vanished AND
-        // whose unpacked layer tree is also gone — genuine cache loss. The run
-        // fails with an actionable re-pull instruction, not the old bare
-        // "rootfs is missing" bail.
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        let mut image = sample_image("docker.io/library/alpine:3.20", digest, "blobs/a");
-        image.rootfs_path = Some("rootfs/alpine/rootfs.ext4".to_string());
-        image.runtime_tag = Some(oci_runtime_tag());
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-        // Deliberately write neither the rootfs.ext4 nor the unpacked tree.
-
-        let err = resolve_or_pull_run_image(tmp.path(), "docker.io/library/alpine:3.20", false)
-            .expect_err("missing rootfs with no unpacked layers must fail");
-        let msg = format!("{err:#}");
-        assert!(
-            msg.contains("mvmctl image pull"),
-            "error should tell the user to re-pull: {msg}"
-        );
-        assert!(
-            msg.contains("unpacked"),
-            "error should explain the unpacked layers are gone: {msg}"
-        );
-    }
-
-    #[test]
-    fn stale_runtime_tag_marks_rootfs_not_current() {
-        let mut image = sample_image("docker.io/library/alpine:3.20", "sha256:dead", "blobs/a");
-        image.rootfs_path = Some("rootfs/alpine/rootfs.ext4".to_string());
-        let current = oci_runtime_tag();
-
-        // Pre-tag entry (None): the baked agent predates the tag, so stale.
-        assert!(!cached_rootfs_is_current(&image, &current));
-
-        // A different epoch/version means a different injected runtime — stale.
-        image.runtime_tag = Some("0.0.0.0".to_string());
-        assert!(!cached_rootfs_is_current(&image, &current));
-
-        // Matching tag is the only current case.
-        image.runtime_tag = Some(current.clone());
-        assert!(cached_rootfs_is_current(&image, &current));
-
-        // A current tag with no materialized rootfs is still not bootable.
-        image.rootfs_path = None;
-        assert!(!cached_rootfs_is_current(&image, &current));
-    }
-
-    #[test]
-    fn rootfs_verity_sidecars_present_requires_both_files() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let rootfs = tmp.path().join("rootfs.ext4");
-        std::fs::write(&rootfs, b"rootfs").expect("write rootfs");
-        assert!(!rootfs_verity_sidecars_present(&rootfs));
-
-        std::fs::write(tmp.path().join("rootfs.verity"), b"verity").expect("write verity");
-        assert!(!rootfs_verity_sidecars_present(&rootfs));
-
-        std::fs::write(tmp.path().join("rootfs.roothash"), b"abc\n").expect("write roothash");
-        assert!(rootfs_verity_sidecars_present(&rootfs));
-    }
-
-    #[cfg(feature = "pure-mkfs")]
-    #[test]
-    fn resolve_run_image_reseals_cached_rootfs_when_verity_sidecars_are_missing() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let digest = "sha256:bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb";
-        let mut image = sample_image("docker.io/library/alpine:3.20", digest, "blobs/a");
-        image.rootfs_path = Some("rootfs/alpine/rootfs.ext4".to_string());
-        image.runtime_tag = Some(oci_runtime_tag());
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-        write_minimal_config(tmp.path());
-        write_file(tmp.path(), "rootfs/alpine/rootfs.ext4", b"stale-rootfs");
-        let unpacked = tmp
-            .path()
-            .join("unpacked")
-            .join(sha256_hex(digest).expect("hex digest key"));
-        std::fs::create_dir_all(unpacked.join("etc")).expect("create unpacked tree");
-        std::fs::write(unpacked.join("etc/hostname"), b"box\n").expect("write unpacked file");
-        seed_guest_runtime_cache(tmp.path());
-
-        let resolved =
-            resolve_or_pull_run_image(tmp.path(), "docker.io/library/alpine:3.20", false)
-                .expect("stale verity-free cached rootfs must be re-sealed");
-
-        assert!(resolved.rootfs_path.ends_with("rootfs/alpine/rootfs.ext4"));
-        assert!(resolved.rootfs_path.is_file());
-        assert!(
-            resolved
-                .rootfs_path
-                .parent()
-                .unwrap()
-                .join("rootfs.verity")
-                .is_file()
-        );
-        assert!(
-            resolved
-                .rootfs_path
-                .parent()
-                .unwrap()
-                .join("rootfs.roothash")
-                .is_file()
-        );
-    }
-
-    #[test]
-    fn provenance_labels_cover_claim_10_fields() {
-        let image = sample_image(
-            "docker.io/library/alpine@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            "blobs/a",
-        );
-
-        let trust = OciTrustDecision {
-            trust_policy: "prod-cosign-required".to_string(),
-            verification_status: "cosign-verified identity=repo issuer=issuer".to_string(),
-        };
-        let provenance = image.provenance("image_pull", "alpine@sha256:aaa", &trust);
-        let labels: BTreeMap<_, _> = provenance.audit_labels().into_iter().collect();
-
-        assert_eq!(provenance.registry, "docker.io");
-        assert_eq!(provenance.repository, "library/alpine");
-        assert_eq!(
-            provenance.resolved_digest,
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-        );
-        assert_eq!(provenance.layer_digests, vec!["sha256:layer"]);
-        assert_eq!(provenance.trust_policy, "prod-cosign-required");
-        assert_eq!(
-            labels.get("oci_supplied_reference").map(String::as_str),
-            Some("alpine@sha256:aaa")
-        );
-        assert_eq!(
-            labels.get("oci_verification_status").map(String::as_str),
-            Some("cosign-verified identity=repo issuer=issuer")
-        );
-    }
-
-    fn policy_text() -> &'static str {
-        r#"
-allowed_registries = ["docker.io", "ghcr.io"]
-require_signatures = true
-
-[[cosign]]
-certificate_identity = "https://github.com/tinylabscom/mvm/.github/workflows/release.yml@refs/tags/v0.14.0"
-certificate_oidc_issuer = "https://token.actions.githubusercontent.com"
-"#
-    }
-
-    #[test]
-    fn oci_policy_parses_registry_allowlist_and_cosign_identity() {
-        let policy = parse_oci_registry_policy(policy_text()).expect("policy parses");
-
-        assert_eq!(policy.allowed_registries, vec!["docker.io", "ghcr.io"]);
-        assert!(policy.require_signatures);
-        assert_eq!(policy.cosign.len(), 1);
-        assert_eq!(
-            policy.cosign[0].certificate_oidc_issuer,
-            "https://token.actions.githubusercontent.com"
-        );
-    }
-
-    #[test]
-    fn production_policy_accepts_valid_cosign_signature() {
-        let policy = parse_oci_registry_policy(policy_text()).expect("policy parses");
-        let image_ref: ImageReference = "docker.io/library/alpine@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-            .parse()
-            .expect("valid image ref");
-        let verifier = MockCosignVerifier::new(vec![Ok(())]);
-
-        let trust = enforce_oci_trust_policy_with(
-            &image_ref,
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            &policy,
-            &verifier,
-        )
-        .expect("valid signature accepted");
-
-        assert_eq!(trust.trust_policy, "prod-cosign-required");
-        assert!(trust.verification_status.contains("cosign-verified"));
-    }
-
-    #[test]
-    fn production_policy_rejects_missing_signature() {
-        let policy = parse_oci_registry_policy(policy_text()).expect("policy parses");
-        let image_ref: ImageReference = "docker.io/library/alpine@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-            .parse()
-            .expect("valid image ref");
-        let verifier = MockCosignVerifier::new(vec![Err(CosignVerifyError::MissingSignature(
-            "no matching signatures".to_string(),
-        ))]);
-
-        let err = enforce_oci_trust_policy_with(
-            &image_ref,
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            &policy,
-            &verifier,
-        )
-        .expect_err("missing signature rejected");
-
-        assert!(err.to_string().contains("missing signature"));
-    }
-
-    #[test]
-    fn production_policy_rejects_invalid_signature() {
-        let policy = parse_oci_registry_policy(policy_text()).expect("policy parses");
-        let image_ref: ImageReference = "docker.io/library/alpine@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-            .parse()
-            .expect("valid image ref");
-        let verifier = MockCosignVerifier::new(vec![Err(CosignVerifyError::InvalidSignature(
-            "certificate identity mismatch".to_string(),
-        ))]);
-
-        let err = enforce_oci_trust_policy_with(
-            &image_ref,
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            &policy,
-            &verifier,
-        )
-        .expect_err("invalid signature rejected");
-
-        assert!(err.to_string().contains("invalid signature"));
-    }
-
-    #[test]
-    fn production_policy_rejects_denied_registry_before_cosign() {
-        let policy = parse_oci_registry_policy(policy_text()).expect("policy parses");
-        let image_ref: ImageReference = "quay.io/acme/app@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-            .parse()
-            .expect("valid image ref");
-        let verifier = MockCosignVerifier::new(vec![Ok(())]);
-
-        let err = enforce_oci_trust_policy_with(
-            &image_ref,
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            &policy,
-            &verifier,
-        )
-        .expect_err("registry denial rejected");
-
-        assert!(err.to_string().contains("denied by production policy"));
-        assert!(verifier.results.borrow().len() == 1, "cosign must not run");
-    }
-
-    #[test]
-    fn production_policy_requires_trusted_identity_when_signatures_required() {
-        let err = parse_oci_registry_policy("allowed_registries = [\"docker.io\"]\n")
-            .expect_err("missing identity rejected");
-
-        assert!(err.to_string().contains("no [[cosign]] trusted identity"));
-    }
-
-    #[test]
-    fn production_policy_rejects_signature_opt_out() {
-        let err = parse_oci_registry_policy(
-            r#"
-allowed_registries = ["docker.io"]
-require_signatures = false
-"#,
-        )
-        .expect_err("prod policy cannot opt out of signatures");
-
-        assert!(
-            err.to_string().contains("cannot disable cosign signatures"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn registry_env_key_normalizes_registry_host() {
-        assert_eq!(trust::registry_env_key("ghcr.io").expect("key"), "GHCR_IO");
-        assert_eq!(
-            trust::registry_env_key("registry.example.test:5000").expect("key"),
-            "REGISTRY_EXAMPLE_TEST_5000"
-        );
-    }
-
-    #[test]
-    fn registry_auth_prefers_registry_specific_bearer_token() {
-        let image_ref: ImageReference = "ghcr.io/acme/app:latest".parse().expect("image ref");
-        let auth = trust::registry_auth_from_lookup(&image_ref, |name| match name {
-            "MVM_OCI_BEARER_TOKEN_GHCR_IO" => Some("registry-token".to_string()),
-            "MVM_OCI_BEARER_TOKEN" => Some("global-token".to_string()),
-            _ => None,
-        })
-        .expect("auth resolution");
-
-        assert_eq!(auth.source, "env:MVM_OCI_BEARER_TOKEN_GHCR_IO");
-        assert!(auth.auth.is_authenticated());
-        assert_eq!(auth.auth.kind(), "bearer");
-    }
-
-    #[test]
-    fn registry_auth_falls_back_to_global_bearer_token() {
-        let image_ref: ImageReference = "ghcr.io/acme/app:latest".parse().expect("image ref");
-        let auth = trust::registry_auth_from_lookup(&image_ref, |name| match name {
-            "MVM_OCI_BEARER_TOKEN" => Some("global-token".to_string()),
-            _ => None,
-        })
-        .expect("auth resolution");
-
-        assert_eq!(auth.source, "env:MVM_OCI_BEARER_TOKEN");
-        assert_eq!(auth.auth.kind(), "bearer");
-    }
-
-    #[test]
-    fn registry_auth_has_no_docker_config_dependency() {
-        let image_ref: ImageReference = "ghcr.io/acme/app:latest".parse().expect("image ref");
-        let requested = RefCell::new(Vec::new());
-        let auth = trust::registry_auth_from_lookup(&image_ref, |name| {
-            requested.borrow_mut().push(name.to_string());
-            None
-        })
-        .expect("auth resolution");
-
-        assert_eq!(auth.source, "anonymous");
-        assert!(!auth.auth.is_authenticated());
-        assert_eq!(
-            requested.into_inner(),
-            vec![
-                "MVM_OCI_BEARER_TOKEN_GHCR_IO".to_string(),
-                "MVM_OCI_BEARER_TOKEN".to_string()
-            ]
-        );
-    }
-
-    #[test]
-    fn manifest_config_descriptor_extracts_config_blob() {
-        let digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        let manifest = serde_json::json!({
-            "schemaVersion": 2,
-            "config": {
-                "mediaType": "application/vnd.oci.image.config.v1+json",
-                "digest": digest,
-                "size": 17,
-            },
-            "layers": [],
-        });
-        let descriptor =
-            manifest_config_descriptor(&serde_json::to_vec(&manifest).unwrap()).unwrap();
-        let descriptor = descriptor.expect("config descriptor");
-        assert_eq!(descriptor.digest, digest);
-        assert_eq!(descriptor.size, 17);
-    }
-
-    #[test]
-    fn upsert_replaces_existing_reference_entry() {
-        let mut index = OciCacheIndex {
-            schema_version: 1,
-            images: vec![sample_image(
-                "docker.io/library/alpine:3.20",
-                "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "blobs/old",
-            )],
-        };
-        let replacement = sample_image(
-            "docker.io/library/alpine:3.20",
-            "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            "blobs/new",
-        );
-
-        upsert_image(&mut index, replacement);
-
-        assert_eq!(index.images.len(), 1);
-        assert_eq!(index.images[0].layers[0].path.as_deref(), Some("blobs/new"));
-    }
-
-    #[test]
-    fn upsert_keeps_distinct_references_to_same_digest() {
-        let digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        let mut index = OciCacheIndex {
-            schema_version: 1,
-            images: vec![sample_image(
-                "docker.io/library/alpine:3.20",
-                digest,
-                "blobs/shared",
-            )],
-        };
-        let second = sample_image(
-            &format!("docker.io/library/alpine@{digest}"),
-            digest,
-            "blobs/shared",
-        );
-
-        upsert_image(&mut index, second);
-
-        assert_eq!(index.images.len(), 2);
-    }
-
-    #[test]
-    fn missing_index_lists_empty() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let rows = list_rows(tmp.path(), None).expect("list");
-        assert!(rows.is_empty());
-    }
-
-    #[test]
-    fn registry_filter_limits_list_rows() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let mut other = sample_image("ghcr.io/acme/app:1", "sha256:app", "blobs/app");
-        other.registry = "ghcr.io".to_string();
-        other.repository = "acme/app".to_string();
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![
-                    sample_image("docker.io/library/alpine:3.20", "sha256:alpine", "blobs/a"),
-                    other,
-                ],
-            },
-        );
-
-        let rows = list_rows(tmp.path(), Some("ghcr.io")).expect("list");
-        assert_eq!(rows.len(), 1);
-        assert_eq!(rows[0].reference, "ghcr.io/acme/app:1");
-    }
-
-    #[test]
-    fn inspect_resolves_by_reference_and_digest() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let image = sample_image("docker.io/library/alpine:3.20", "sha256:alpine", "blobs/a");
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-        write_file(
-            tmp.path(),
-            "manifests/alpine.json",
-            br#"{"schemaVersion":2}"#,
-        );
-        write_file(
-            tmp.path(),
-            "configs/alpine.json",
-            br#"{"architecture":"arm64"}"#,
-        );
-        write_file(
-            tmp.path(),
-            "claims/alpine.json",
-            br#"{"labels":{"mvm":"yes"}}"#,
-        );
-
-        let by_ref =
-            inspect_image(tmp.path(), "docker.io/library/alpine:3.20").expect("inspect by ref");
-        let by_digest = inspect_image(tmp.path(), "alpine").expect("inspect by short digest");
-        assert_eq!(by_ref.image.reference, by_digest.image.reference);
-        assert!(by_ref.manifest.is_some());
-        assert!(by_ref.config.is_some());
-        assert!(by_ref.claims.is_some());
-    }
-
-    #[test]
-    fn remove_refuses_paths_that_escape_cache_root() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let mut image = sample_image("docker.io/library/alpine:3.20", "sha256:alpine", "../bad");
-        image.manifest_path = "../manifest.json".to_string();
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![image],
-            },
-        );
-
-        let err = remove_image(tmp.path(), "sha256:alpine").expect_err("unsafe path rejected");
-        assert!(err.to_string().contains("escapes cache root"));
-        let index = load_index(tmp.path()).expect("index still readable");
-        assert_eq!(index.images.len(), 1);
-    }
-
-    #[test]
-    fn remove_preserves_shared_layers_and_updates_index() {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let first = sample_image(
-            "docker.io/library/alpine:3.20",
-            "sha256:first",
-            "blobs/shared",
-        );
-        let mut second = sample_image(
-            "docker.io/library/busybox:1",
-            "sha256:second",
-            "blobs/shared",
-        );
-        second.manifest_path = "manifests/busybox.json".to_string();
-        second.config_path = None;
-        second.claims_path = None;
-        write_index(
-            tmp.path(),
-            &OciCacheIndex {
-                schema_version: 1,
-                images: vec![first, second],
-            },
-        );
-        write_file(tmp.path(), "manifests/alpine.json", b"{}");
-        write_file(tmp.path(), "configs/alpine.json", b"{}");
-        write_file(tmp.path(), "claims/alpine.json", b"{}");
-        write_file(tmp.path(), "blobs/shared", b"layer");
-
-        let outcome = remove_image(tmp.path(), "sha256:first").expect("remove");
-        assert_eq!(outcome.removed_files, 3);
-        assert!(tmp.path().join("blobs/shared").exists());
-        let index = load_index(tmp.path()).expect("load index");
-        assert_eq!(index.images.len(), 1);
-        assert_eq!(index.images[0].reference, "docker.io/library/busybox:1");
-    }
 }
